[roll] Roll fuchsia [hlcpp][gidl] In HLCPP conformance test, don't fake out transaction header
Motivation:
- This adds a lot of complication to the test which makes it harder
to understand and verify that it is right.
- The decode failure test was actually not working - it was catching
failures due to the wrong FIDL type.
- Need to update the coding table for the transaction header manually
when coding tables change (this is actually my original motivation
because a change I'm working on made it complex to build the coding
table for this manually).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407154
Original-Revision: d824ec756f0dacd3f7a255f4ab27a6ef54385c20
GitOrigin-RevId: a53ec42538710730605f369420cb57f1a7449122
Change-Id: I2dac7f8b1cfe91aaeb52815d2be45d1730903013