[roll] Roll fuchsia [storage] Fix the client so that it sends CloseFifo when dropped.
Also fixed a bug in the block server where closing the fifo could get
blocked if unable to return responses on the fifo (which can be the case
if the client isn't servicing the fifo).

Changed termination for the Rust client so that the fifo is actively
closed.

Original-Bug: 54733
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/404198
Original-Revision: 9880d7f9a24077fa4d591c873b428cf2f6d6b792
GitOrigin-RevId: 64018543f70999488101e2a9514ac5a2c6424754
Change-Id: I9a3a7589515abb77f62444c084de4ef629d03d9a