[roll] Roll fuchsia [scrutiny] New dashboard landing page.
This change adds a new landing page for scrutiny with a button that you
click to move to the graph. This enables us to add more visualizers in
the future with a uniformed landing dashboard. This is just a very
incremental change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407211
Original-Revision: e389bcfdfa944f941dcacbc431c3a813624b0da5
GitOrigin-RevId: df849437f0a7b1666522ee170d1dc8e925bfde43
Change-Id: Ieee637a915ffb9650107206a65c6cd6c18a5a909