[roll] Roll fuchsia [scripts][fuzzing] Handle case where package is in TUF repo but not on disk
`pkgctl pkg-status` returns an exit code of 2 when a package is built as
a universe package but not yet resolved to disk. This was being
interpreted as an error, preventing package resolution from proceeding
as expected. This CL ensures that such an exit code is handled properly,
and extends the process mocking framework to handle variable exit codes
so that the behavior can be tested.

Original-Fixed: 55235
Test: fx fuzz unittest
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/406685
Original-Revision: d16427eec478a0279266094621269009fee74513
GitOrigin-RevId: ae4ff05c8a2cadbd255190e449ab206e463937af
Change-Id: I317452744572c5c5cb7afada078ff213a906189d