[roll] Roll fuchsia [settings] Add transitional tags to Watch()
There are some fakes that depend on an exact match to the settings api
interfaces, so we will need the transitional tag to generate bindings
for those that do not implement it yet. This will be removed later once
the migration is complete.

Original-Fixed: 56309
Original-Bug: 56309, b/161125222
Test: fx test setui_service_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407198
Original-Revision: f841c8eb948c999ac93b0d063ccf9aeb82bb68c7
GitOrigin-RevId: ca9ebbba1a3cd3390da95551a208b9b869acbf31
Change-Id: I70e8ed3aef7ee8a3e0674a2c65f7877b140af51b