[roll] Roll fuchsia [debugger] Remove the arch provider
Moves the breakpoint and watchpoint functions to the DebugRegisters
object. The DebugRegisters implementation is now split off into an x64
and an arm64 implementation.

Move WatchpointInstallationResult to WatchpointInfo in its own file.
This is now used for both installs and queries of the DebugRegisters.
The "status" is removed and failure is indicated (when needed) by
returning in a std::optional.

Remove some zx_status_t results from affected functions. Booleans and
std::optional make the API cleaner, and all we were ever doing with the
results was printing the errors, sometimes multiple times. Now we print
it once and just pass a boolean up the stack.

The breakpoint instruction checking function is split out into an
architecture-specific version that checks a passed-in value, and a
variant on the DebuggedThread that actually fetches the memory.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407283
Original-Revision: adf40b3757eb21474bad46ad85b856a1936cff76
GitOrigin-RevId: e16fd63ab03e131023c46216f338c2bcb27ddcb8
Change-Id: I3e6b43f083b80c5cca50df43933a9db4bb8ffa95