[roll] Roll fuchsia [fragment][driver] Add SCPI protocol client
The fragment driver currently does not support the SCPI protocol. The
result is that fragments that are bound to a SCPI device end up with the
protocol ZX_PROTOCOL_MISC. This leads to a situation where a composite
device which contains SCPI fragment will be unable to access that
fragment.

Original-Fixed: 56284
Test: run device-enumeration-test locally on VIM2
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407278
Original-Revision: 0222ee38cb873b9c52928b11805dcad80dc80f9b
GitOrigin-RevId: 09d19ab4e0fe299c04e369c42a66950481e92e1d
Change-Id: I1cb3492818d409288045bb4a71bb4bed596a3f89