[roll] Roll fuchsia [feedback] Add test FIDL to query CrashReporter
CrashReporter, like Cobalt, is a component that is sent data and
is often ignored when it responds. This make integration tests
challenging because a test may not fail if something is wrong with
a component's crash reporting logic and a crash report is not filed.
Adding a query interface allows component authors to easily check
that crash reports are filed in integration tests.

Original-Bug: 55875

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407118
Original-Revision: b5f3c8aa16e5e3c7318ca416547786fe90023eb3
GitOrigin-RevId: 5fd0246d4ade1114b5e6518e6cddc4788c027147
Change-Id: I5a10773709178c143cf4c04aaccda928d3bffdb4