[roll] Roll fuchsia [bt][a2dp] Refactor RtpPacketBuilder into library
Delegate the building of the RtpPacketBuilder to the MediaCodecConfig,
as it holds all the information to build it.

Introduce the RtpPacketBuilder trait which will be used in the future
to build more codec-specific behavior for RTP packet builders.

Original-Bug: 52616

Test: refactor, tests moved to bt-a2dp-tests

Multiply: bt-a2dp-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/406319
Original-Revision: 79dbe76e390bc70d6ef3425e286d50a78943138c
GitOrigin-RevId: d84f95710815c0236f57334644059d277bd6709a
Change-Id: Ifc042259145ed36e34a841a4fa373fcc98121234