[roll] Roll fuchsia [bt][avrcp-tg] Fix log spam from notifications
In AVRCP, notifications are a one-use object. When the notification
value changes, the changed value is sent over the responder, and the
notification object is dropped. However, there are some cases where the
notification value never changes. In these cases, when the Notification
object is dropped, the current value is sent over the responder. The
sending of this value was being logged, and in some cases, the send
fails - this is OK. It usually means the client stopped caring about the
notification value or the client disconnected. This CL removes any
logging around the send operation as this was causing log spam.

While we're here, update the component to use log crate macros - this is
the standard guidance to avoid mismatches in severities across
components.

Original-Bug: 56335
Test: Log change, fx test bt-avrcp-target-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407616
Original-Revision: decae920a5db3774fd4e90bc7e673fd90359e2d6
GitOrigin-RevId: 87027223c829f72e7db10d5e27499c0b57e3f1e9
Change-Id: Ib55c5a63e810a76e8a29eafac38c78c254d7fe42