[roll] Roll fuchsia [media][audio_core] Apply OutputPipeline Updates
To continue enabling the AudioTuner service, this CL implements the
proper update of the OutputPipeline with a provided PipelineConfig and
VolumeCurve.

Two key factors to proper replacement of the OutputPipeline are:
	> Maintenance of input streams
	> Execution on the mix thread

Input streams are maintained by forcing the re-route of these inputs
through the new pipeline by removing the output device from the route
graph and adding it following the pipeline update; this takes advantage
of existing mechanisms for routing streams upon the addition of a
device. Because of this, the AudioTuner method to replace the
OutputPipeline is only valid on a device without links.

The actual replacement of the OutputPipeline is executed on the mix
thread so as to align with existing mix tasks and provide a smooth
transition between OutputPipeline configurations. Executing on the mix
thread ensures the application of the new pipeline occurs between mixes,
never during a mix. A 'routable' boolean is used to protect the
OutputPipeline during update; 'routable' is false during update such
that the device plug state cannot change and another update cannot be
simultaneously applied; 'routable' is made true upon update of the
PipelineConfig.

Original-Bug: 52962

Test: fx test audio_pipeline_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/398960
Original-Revision: ff25c345770e40b8016df1f29bbe5d01eafc15c4
GitOrigin-RevId: 36b63a7a35a6284c549a51cd05613abb8f739379
Change-Id: I23656b033d986294ceef161bf61e6f093e7269a3