[roll] Roll fuchsia [bt][a2dp] Migrate AvdtpController to use a2dp::Peer
A2dpController (renamed from AvdtpController) now operates on an
a2dp::Peer, allowing for manipulation of the underlying avdtp::Peer and
the maintanence of state.

Some PTS tests required stateful behavior (i.e
out of band suspend results in the command being sent AND the underlying
Stream being in the right state). Using the a2dp::Peer, which maintains
the collection of Streams, the state of the peer, and the avdtp::Peer
allows us to achieve this.

Updated logging to use log as opposed to fuchsia_syslog.

Updated the A2dpControllerPool to use the already-discovered streams
from the a2dp::Peer, as opposed to attempting to discover when needed.
This was confusing PTS as stray avdtp.discover() commands were being
sent.

Original-Bug: 56032
Test: fx test bt-a2dp-tests, PTS AVDTP/SRC/INT/SIG/SMG/BV-21-C

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/406059
Original-Revision: 7557b5bf72ff186df2b181fb3537880ee2281cfc
GitOrigin-RevId: 2b1c5857696218888aab0d2060556ecbc0ab3b52
Change-Id: Idb232fb164a0abb8a9ac379371af58c59f0e3891