[roll] Roll fuchsia [fx, wfh] warn users about unsupported wfh flows
In the serve-remote flow it is not possible to adjust device targets
on the remote side. We detect the remote side by looking for the
forwarded target address, and print a warning in both `fx set-device`
and `fx -d` use cases.

Original-Fixed: 49685
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/398278
Original-Revision: 1f4caef21f622d30258c143929df90c5045a90a9
GitOrigin-RevId: b592e7255fa5d6426b5a4b8c90edb88047b36eee
Change-Id: Ic1f0835c58ef35fc128aa20b35e04b333d388353