[roll] Roll fuchsia [blobfs] Fix race in unmount
Currently, async::Loop requires that we only add or remove paged VMOs
from its internal linked list serially.

Blobfs violates this invariant during teardown. The pager object is
destroyed in the main thread, which iterates through and cancels all of
the paged VMOs in the async::Loop. But it is possible for another thread
(in particular, the journal thread) to simultaneously modify the list,
by deleting a Blob object, which unregisters a paged VMO from the
async::Loop.

This fix is a workaround that makes the race condition impossible, by
ensuring that the journal operations finish before deleting the pager
object. It is a brittle fix and more attention should be given to how to
properly fix issues of this class.

Original-Bug: 55829
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407599
Original-Revision: 0cdd4e45514b1b585d44787ec38c4806c6d6acbb
GitOrigin-RevId: 4ae2de8134cbf6c0f102794295739c7495db4faf
Change-Id: I9c523260ec328958dc059ab3903691b6ac0c2613