[roll] Roll fuchsia [ninja] Make go/build.py use py3
This is necessary as otherwise when creating the symlink tree, Python 2
will truncate the mtimes of the symlinks, resulting in
older-than-expected mtimes. This is fixed in Py3, and it is hoped this
will resolve the fail no-op as in
https://ci.chromium.org/raw/build/logs.chromium.org/fuchsia/led/fuchsia-try-builder_fuchsia-infra.iam.gserviceaccount.com/ef71493e6e5c8e1586c2d2af0a34f3ec604c4ba2eb663d90e34a591ec361b438/+/build.proto?server=chromium-swarm.appspot.com

    """
    ninja explain: output host_x64/gen/tools/testing/affectedtests/testdata older than most recent input ../../tools/testing/affectedtests/testdata (1594856370119459000 vs 1594856370119459941)
    """

In particular note that the dates are the same except the last three
digits which have truncated nanoseconds to microseconds.

Original-Bug: 56376
Original-Bug: 36581
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407784
Original-Revision: 8643f1025c83f1282d5cb21fc4dc40588cc4b603
GitOrigin-RevId: 0f14788fed9d709fc2b078fba08760e398a98c11
Change-Id: I90c76a843e23f6142cad307ac2c9a83d862c741c