[roll] Roll fuchsia [bsdsocket_test] Avoid clients connect/accept race.
On Linux, ReadBeforeConnect tests with multiple client connects to the
server and intends to accept the connections in the order of connect
calls. The existing test has a race where the client connects can get
submitted to the TCP stack's server accept queue out of order. To fix
this, we ensure that the test observes the completed precursor client
connection on the server side before attempting subsequent client
connects.

Original-Fixed: 55642
Test: fx test --host -o netstack_bsdsocket_c_api_test -- --gtest_filter=NetStreamTest.ReadBeforeConnect --gtest_repeat=1000

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407895
Original-Revision: e925deca8204c1486a6430e43070f6d0ad973277
GitOrigin-RevId: 4dfbf41b04d15e3bcef86cfb9c54ec083b21a178
Change-Id: Ieb0942191a223e2e8cf3400aee6bbd11fd206de7