[roll] Roll fuchsia [ninja][build] Ensure there's always a .sym output
When there's no buildid for the kernel, a special case to handle
dump_syms results in skipping emitting the .sym file (by not running
dump_syms). This causes the build to always be dirty because the
'outputs' still specify that the .sym will be output. To avoid this,
touch to make an empty file.

Original-Bug: 56377
Original-Bug: 56374
Original-Bug: 36581
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407777
Original-Revision: 09de4ce370d3a9db913ec771e90c9215c8ccfcf0
GitOrigin-RevId: 720061e560e90a2a20d455987f48b05fe20c92db
Change-Id: Id6e31b90e3d19c4b31760031a9715d2782087841