[roll] Roll fuchsia [nvme] Migrate from DEVICE_ADD_INVISIBLE to init hook.
The device init hook is called after device_add().

If the init hook is implemented, the device is guaranteed
to be invisible and unable to be unbound until
device_init_reply() is called.

If device_init_reply() is passed an error, the device
will be scheduled to be unbound.

BUG=43261

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/406652
Original-Revision: 2705d527d4302a98224f5acf79bb5e8fba3e17dd
GitOrigin-RevId: d15e0ad4a0ab3a4046df690fab9692e0eeea0b80
Change-Id: I151d1c5c8ec9b2823b048d30a4d038ad1eefb1f0