[roll] Roll fuchsia [bti][pmt] Warn when pinning with a BTIs which has quarantined pages.
If a driver developer attempts to pin memory while a BTI has
quarantined pages, print a warning, and prepare for the day that we
will deny the request instead.

While there are a few different reasons that pages might end up
getting quarantined, a common reason is that a driver crashed and was
re-loaded.  If the driver author forgets that they need to always
assume nothing, and make sure that their DMA is stopped before
starting to operate, then they probably forgot to call release
quarantine as well.  By denying new pin operations, hopefully the
developer will become aware of the page quarantine protocol and fix
their code properly instead of just continuing to permanently leak
pages forever.

Right now, we are not ready to simply deny the request as this could
lead to a lot of fallout in CQ.  For now, we just warn instead, but
the CL includes all of the code and tests needed to eventually deny
the request instead.

Original-Bug: 56205
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407129
Original-Revision: bee1fbf7f62748c3948af377762411be4b43b953
GitOrigin-RevId: 915fe3ecc7698765ea8783a4c31526834715bb65
Change-Id: I5ee136ac22b8a7ba9bb49f7e0b2b125575b659a9