[roll] Roll fuchsia [audio_core] AudioClock refinement
AudioClock objects are passed by reference.These (and their
underlying zx::clock objects) are owned by either an AudioDriver,
BaseRenderer or BaseCapturer, so we can guarantee that the
ClockReference object will outlive those that are given a
reference.

The zx::clock object at the core of AudioClock is now entirely
encapsulated, as shown by the deletion of the clock() accessor
and the addition of utility methods that previously would have
required the raw zx::clock.

AudioClock is updated to include a source (client, device) and a
type (adjustable, non-adjustable, invalid) as well as a clock
domain and a flag that represents whether a client clock has
been designated as controlling the clock hardware.
of 'adjustable' is dropped.

Also, MixStage::SetupMix is reorganized to prepare for subsequent
CL that adds clock reconciliation.

Test: 2 new AudioClock unittests, >50 existing tests adjusted
Original-Bug: 46601

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405594
Original-Revision: 831d50e4491646cb82b41d1b0735684660712fb1
GitOrigin-RevId: 4717c2a82e38fc0221130513c06824bf0723eff7
Change-Id: I880e18afc8640d6105b8dbba1189fa79a9e358a5