[roll] Roll fuchsia [archivist] Open matched components' directories only
When reading Inspect snapshots, we should filter down so that we only
read snapshots for components that may possibly contain the Inspect data
we are looking for. Now that we have several components sampling this
data simultaneously (for tests and other reports) this change provides a
drastic improvement in Archivist's memory and CPU usage.

Original-Bug: 55302
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407958
Original-Revision: cdcef6c2f40320fa662559606fb8a5010ad6626f
GitOrigin-RevId: 2243b84810ed003cce522dbd76cf32b3f216a09f
Change-Id: Ied01c46d9d7e0663eb19731f742e73e65b14c5a8