[roll] Roll fuchsia [debugger] Fix failed line lookup.
The issue arises when two functions are exactly adjacent in memory. The
line table will contain an "EndSequence" row for the end of the first
function to mark where the previous line table entry ends. But this is a
non-inclusive end marker and will be at the beginning of the following
function.

Previously, when looking up the line information for the beginning of
the second function, we would find the EndSequence marker for the
previous one.

This happens much more often on ARM because the consistent-length
functions always pick up at the next instruction. x86 code often has
padding which will hide this bug.

This also clarifies (but does not change) the behavior of the
LargestLessOrEqual helper function when it encounters a sequence of
equal items, and and adds a test for this case.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407964
Original-Revision: ab8b7b92ebb1c087f571c86eb5b7e48e1134e133
GitOrigin-RevId: 4a85c8dd261a48cab431c7ec4f621ea3282308ff
Change-Id: I07b8c8f29513256cbf74ed98d91806669578c2ac