[roll] Roll fuchsia [exceptions] Keep exception alive for longer
The exception needs to be kept alive until after the CrashReporter
responds in the event that sysmgr was the process in an exception. Once
sysmgr's exception is handled, all jobs below it, including the Forensics
components will be killed, and the crash report for sysmgr may be
forever lost.

Original-Fixed: 56488

TESTED-`fx test exception-broker-tests`

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407885
Original-Revision: ee402821afacbb9db143902a55eb74a2852100c5
GitOrigin-RevId: c776678297edceb1a115ead97e435a95065e8f84
Change-Id: I84788d1c4274f2cdc9be712b4b9af354ad463643