[roll] Roll fuchsia [kernel][vm] Avoid race in child deletion
This fixes a bug where VmObject referred to in a children_list_ would
be used after it had been destroyed. The use after free occurs as the
children_list_ stores raw pointers, which is mostly fine as it is known
that the child will stay alive until it can acquire the vmo lock. To
avoid a lock ordering inversion OnUserChildRemoved drops the vmo lock
prior to using the object, which is invalid as the 'this' pointer was
a raw pointer from the children_list_.

The solution here is to try and hold a refptr to the child, and only
if a refptr can be upgraded perform the OnUserChildRemoved step so
that the lock can be safely released.

Original-Bug: 56137

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/406757
Original-Revision: 8fbbd5fa8dbcff443a10f0b81b36829965435ce3
GitOrigin-RevId: 8f3e4de4ff0e06dcc47d848545a56394dbd6ce17
Change-Id: Ibc708fd00e4a5c1782965ba3be83614b7c337639