[roll] Roll fuchsia [fs_test] Fix AttrTest.WriteSetsModificationTime
Previously, the test would only assert that the modification time of the
file hadn't gone backwards. Fix this by sleeping for the timestamp
granularity of the filesystem and then writing to it, which means we can
assert that the modification time has actually moved forward.

Also add a test that the modification time is updated even if the file
is not closed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407516
Original-Revision: d9966fef9a51b938579f78fd78f2eda12535bc32
GitOrigin-RevId: d2acc876d2f19ae0e2443e1292ddd85eceb8eab3
Change-Id: I0ea7ddcb44b89929165ab5d2699959378dc7e1a1