[roll] Roll fuchsia [components] Make event types more consistent and improve tests
Prior to this patch, there were two types of v2 component events: ones
that had payloads and ones that didn't. The ones that didn't had a
different test API than the ones that had payloads. That added a bit to
the confusion of using these events. With this CL, all events have
payloads, if only empty payloads. This allows all events to have a
Result type. This also simplifies the macros in events.rs significantly.

This CL also does a couple of things to improve events integration
tests. It moves the tests in a direction where the test function no
longer has a return value and instead all operations call .unwrap().
This will cause them to crash if the call returns an error. Furthermore,
nested components get syslog and LogSink in order to output the result of the
failed unwrap. These will make debugging tests in the future much easier
as we'll know the exact line number failing and the error description
causing the failure in tests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407202
Original-Revision: d706e68f9cacaa1948092388a5dd1ab7f54729e3
GitOrigin-RevId: 61b1069ec4ca9997139348ec7594b41d5a0659dc
Change-Id: I7f5e0ea46ceb4c362316460c78c2180265ed1d3f