[roll] Roll fuchsia [tefmocheck] Add check for failing to run bugreport
One of the more common unattributed task failure modes in the last few
days is the bugreport ssh command failing with no exit code when
testrunner executes it at the end of a task.

Fixing that might be as simple as adding retries, but without detection
in place we won't be able to tell if a fix actually works, so adding a
tefmocheck check is a necessary first step either way.

Original-Bug: 56651
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/408494
Original-Revision: 34aff98bd530c568a729e94b84ec32120af378c8
GitOrigin-RevId: 16f0ea45d3d760caa48b79f4f905c21e60d19571
Change-Id: I79b446b14e22f8a6e0225cc1289eca259cfb6eea