[roll] Roll fuchsia [paver][vmo] Assert that payload VMO is a clone if pager-backed
Add an assert in PartitionPave to verify that if the payload VMO is
pager-backed, it is also a clone. This guarantees that once its pages
are committed, they will remain in memory and not get evicted. We commit
the payload's pages before writing out a partition to prevent deadlocks
in the block server (due to potential page faults).

Test: Manually verified that `fx update` succeeds with and without
demand paging.

Original-Bug: 48145
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/408057
Original-Revision: 482ccff9a55fd64ff4b19fc61bca3ab943c3d6c0
GitOrigin-RevId: a29598cd8ea3e5900625a2c60cdef42bb7fc01ea
Change-Id: I819aa83b9b20b0609d2598a93bbabbef33497e07