[roll] Roll external/github.com/catapult-project/catapult metrics: Add diagnostics to console error metric
https://crrev.com/c/2302837 adds extra details (message and url) to
ConsoleError trace events. This CL adds these info as diagnositcs to
the console error metric so the metric is easier to debug. Additionally,
the pid of the slice is added as a diagnostic since that's also quite
useful for debugging.

The code safely skips adding these details if we're using an older
chrome version and the extra trace arguments are not available on the
ConsoleError slice.

Original-Bug: chromium:1086962
Original-Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2304259
Original-Revision: ade0fed37fc768d0ee94fa3e4fac15e0d0782346
GitOrigin-RevId: 7bf2bc6923b57e0b1712d4e5c1cf693eab8a4d70
Change-Id: Ife113b53d2ac69d47966b7985a9162a618f7997d