[roll] Roll fuchsia [xhci-rewrite] Migrate from DEVICE_ADD_INVISIBLE to DdkInit.
If DdkInit is implemented, the device is guaranteed to be invisible
and not able to be unbound until InitTxn.Reply() is called. Replying
with an error will lead to the device being unbound.

The fake_ddk will automatically call DdkInit following DdkAdd. I have
moved the BufferFactory argument to the UsbXhci constructor, so the
core test can pass in the fake BufferFactory.

BUG=43261

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/408134
Original-Revision: 9426fbaa046ac82e176e1ca3aad95f21912f0ced
GitOrigin-RevId: 5e2df4f3c72b70427027875c574c56997d2dbe14
Change-Id: Ifa21ad70f4f57a1d56348311f3a8dd3527ddcbb2