[roll] Roll fuchsia [serve-remote] supress set-device warnings
I added a feature to set-device that warns the user about running it
under a remote workflow, but this warning fires on subsequent
invocations of serve-remote, which I did not notice because the
remote was behind the version that was being implemented and tested
locally when the warnings were added.

A simple environment variable flag is used to suppress the behavior
for the invocation in question.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/408095
Original-Revision: 82dd59d500b28a378bfe6135a952335294a5dc50
GitOrigin-RevId: 8b746265e92f6646ab51c50346aad0555416343e
Change-Id: I16cb1b0c973f2d66bac8d09537bce0aaffa2e914