[roll] Roll fuchsia [mdns] remove conflicting publications
This CL introduces a new policy for conflicting publications and
fixes some internal issues with object ownership and lifetime.

b/159977686 was caused by a race that occurred when the client
terminated an existing publication and immediately started a new
publication of the same service type and instance name. At times,
the request for the new publication would arrive and the mdns service
before the existing publication was destroyed, these two operations
being signaled on two different channels. The service would reject
the new publication, because it conflicts with the existing
publication.

The new behavior introduced by this CL is to terminate the existing
publication to make way for the new one. This change is implemented
in mdns_service_impl.cc, and just involves removing the object
representing the old publication from a collection.

There was an additional problem at a lower level (in mdns.cc), which
caused the service to behave incorrectly when a publication was
terminated during the probe phase (when the service is determining
whether the service instance is unique on the LAN). This had to do
with collections maintained in the Mdns class and cleanup that needed
to occur based on various events.

The Mdns class maintains a collection agents that are actively
participating in processing and generation of message traffic. A
previous CL (fxr://403396) added cleanup callbacks to that collection,
primarily to remove agents from additional collections used to
access agents by name. This turned out to be the wrong approach to
cleanup given that the InstanceResponder used for publication was
not added to the primary agents collection until the probe was
complete, meaning that cleanup associated with InstanceResponder
could not use the cleanup mechanism during probing (and, indeed, was
not occurring).

The solution used in this CL is to add the ability to attach an
'OnQuit' cleanup closure to an agent, which is called with the agent
quits (which is always part of removing an agent). This mechanism
doesn't involve the primary agent collection, so it can be used
during the probe phase of publication for InstanceResponder.

NOTE: The behavior change (removing old publications in favor of new
ones) seems like a potential security issue. In fact, there are
security issues with both the new behavior and the old behavior.
fxb/56579 has been filed to address this.

Original-Fixed: b/159977686.

Test: fx test mdns_test
      fx test mdns_netemul_test
      New unit test added. Existing integration test modified to
      detect regressions.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407014
Original-Revision: 70002f056f09258877662ebb4a4f95b69d5f5cb8
GitOrigin-RevId: e714655207411b72ac6acd78330856d46ecc52c1
Change-Id: Ia5a10e0bb4200e808762543de14c17673b741d67