[roll] Roll fuchsia [kernel][mp][thread] Assert when blocking from an inactive CPU
When a CPU blocks, the scheduler selects another thread to run on the
CPU.  If there are no other threads ready to run, the special idle
thread will be selected.  In the general case, a CPU must be marked
active in order to schedule a thread.  Therefore, a precondition of
blocking is that the calling CPU be marked active.

Mutex acquisition and wait queue entry already assert that it's OK to
block (via arch_blocking_disallowed).  However, the active/inactive
state is not taken into account.

This change sets "blocking disallowed" to true when a CPU transitions
from active to inactive and false when transitioning back.

Additionally, this change sets blocking disallowed to true early on in
the boot sequence of secondary CPUs as they are not yet active.
However, the boot CPU continues to run with blocking allowed prior to
becoming active.  It's OK for the boot CPU to run like this because
the boot CPU will become active before any secondary CPUs are brought
online and before the idle thread (or any other thread) begins
execution.

Tested by booting qemu/x64+kvm repeated for several hours.

Original-Bug: 55862
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405896
Original-Revision: ca55bc8c45c2d51c6e1b81213046daf998381840
GitOrigin-RevId: ddc6fefd0a93d55f3df88e463e6081eb53d60a3c
Change-Id: Ic8985abece8af0122d7b726f2cb24db08b49f080