[roll] Roll fuchsia [kernel] Avoid using compare_exchange_strong in a loop.
C++ offers both a strong and weak form of the |compare_exchange| atomic
operation. The weak form may have spurious failures which the strong
form avoids, but on platforms where this occurs (such as ARM), this is
typically achieved by having |compare_exchange_strong| be implemented
inside a loop.

When spinning waiting to acquire a Mutex, we already retry failed loads,
so the spurious failures are handled correctly, so we don't need the
extra loop.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/403693
Original-Revision: 6c00534da959ad136ca9055b4c42901e14020048
GitOrigin-RevId: fdc690f177f081ae92597da284c56c420ac02e64
Change-Id: I286f574e77884a2bdeb1292abbba2b12a2403322