[roll] Roll fuchsia [netstack] Avoid proxy storage race
Store the event proxy before starting the handler goroutine, avoiding a
race in which the handler goroutine starts and exits entirely before the
proxy is stored, resulting in the proxy being stored indefinitely.

Avoid suppressing ErrBadhandle while I'm here, since it should not be
possible for the handle to be invalid before the handler goroutine is
started.

There is still a small race possible between the handler goroutine's
cleanup and an interface change event, but such logs are expected to
appear only once, and we'd like to be made aware if this occurs.

Original-Fixed: 56733
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/408858
Original-Revision: b26bf8885225306ad0fc8d013059b047d5cc2725
GitOrigin-RevId: 57ed8a5cf30a7369707401be14427fa7c56447c1
Change-Id: Iba7338bdb590304d4987c22e44a3d93573b63c93