[roll] Roll fuchsia [syslog][rust] Initialize without double-tagging.
fuchsia-syslog always provides a "target" as a tag which
defaults to the module path where the log macro is invoked.
This duplicates the default tag that the syslog library
initializes from the process name. Since we know that we'll
always be providing our own useful tag, pass an empty list of
tags to the underlying library so that it doesn't initialize
with its default process name tag.

Original-Fixed: 55647
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407999
Original-Revision: 3f3299488c71ff87551c49d5f26e2e088c487dee
GitOrigin-RevId: 8982f836069b8819f9414acedeb4e5a9b7df46ae
Change-Id: I33175552356b89b567faad9b83d8c39e699fbb9b