[roll] Roll fuchsia [bootsvc] Skip unnamed VMOs early
We currently error out late when there's an unnamed VMO passed to us.
Instead, skip the VMO early in processing.

We can receive an unnamed VMO if InstrumentationData::GetVmos() does not
call InstrumentationData::Publish() for a particular VMO. Since we can't
close the handle to the VMO, as there's an expectation of a fixed set of
handles in the userboot message, we must deal with this later in
bootsvc.

Original-Bug: b/159855311
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/408836
Original-Revision: 53fe5e0f1180decf690baf37009541f8ff3dbac5
GitOrigin-RevId: e74954c860e4734ace1a74c3fb5c04eec582b57f
Change-Id: Ic19460de364ba1cc697227275310b6303f6c33fa