[roll] Roll fuchsia [camera] Add linear back-off for sysmem free space probe.
This change adds a linear back-off to the delay with which sysmem is
probed for free space. Previously, sysmem would occasionally not free
deallocated memory within the fixed delay, and so eventually fail to
allocate buffers within the fixed 100ms time. This change increases the
initial delay to 200ms, and increases it each attempt up to a maximum of
1000ms.

Test: Ran camera-gym and observed expected behavior.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409036
Original-Revision: a5cee21a8f8b7c3f449a62f96fe5047b4131c7aa
GitOrigin-RevId: 46b0c301c90923ac46d8544c7caf00d50a9d407c
Change-Id: Icaff2065385a05a7d9aa9be1c6d8393731fcbff5