[roll] Roll fuchsia [build] Add minimal support to build Win32 binaries.
This will be needed to generate the UEFI boot loader, which
must be a Win32 PE32+ binary. This CL only modifies a few
necessary configs when current_os is "win", which none of
the existing toolchain() instances in the Fuchsia build
support anyway.

These changes are here to reflect the content of
//zircon/public/gn/config/BUILD.zircon.gn which provides many
configs that are similar to the ones in
//build/config/BUILD.gn

+ Move //build/config/fuchsia:icf to //build/config:icf
  since this will be used by Zircon-specific toolchains
  as well.

Original-Bug: 54161
Testing: This should not change the build.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/401617
Original-Revision: 5b65af36346f4ac10796e0901d3e621a8d356dbb
GitOrigin-RevId: 49f6648b5a51d3fbeac1a488f29f374fc79201ba
Change-Id: Ic89c923db60321463ef34fbe4fa42d0bfa00ecec