[roll] Roll fuchsia [serial] Add option for server to start at end
As we move to persistent serial logs, we need a multiplexer that
starts new connections at the current point in time. However, the
current behavior is necessary for the tight interval botanist needs
to run in today, so make this an option that can be passed into the
server.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409258
Original-Revision: 26a8332dbf31349badd16fcefc0bcae38caccdd3
GitOrigin-RevId: 3ab3ce1ddacf5dc0e52ed97e2e500332e613c2e4
Change-Id: Ic836e9bb5a144224d320bcbe894933cac1e878dd