[roll] Roll fuchsia [serial] Assume socket server starts at end
Botanist used to have to copy through a large amount of data to get
to the current location in the serial log in the persistent serial case.
In other words, when a socket server was started outside of the task, we
had to process a very large amount of data. We can skip this now that
the serial server supports starting at the end of the aux output.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409261
Original-Revision: d95382d18aa70a25ce22302e6474da5f54350ef9
GitOrigin-RevId: bcc6e350152801f4f993a881554d65659f1a5591
Change-Id: I94a8b76ebeeff5ebbe875bb999e02746231fdd9d