[roll] Roll fuchsia [temperature] Add TMP112 driver
Adds a new driver for the TI TMP112 temperature sensor connected over
I2C. There are not currently any boards that this driver will match on.

To validate the new driver I connected a TMP112 sensor to a VIM2 and
modified the VIM2 I2C devices to allow this driver to bind. The driver
was bound as expected and was verified to dispatch valid temperature
samples using thermal-cli. Some results are documented in fxbug.dev/56484.

Test: fx test tmp112-test
Test: Local validation using TMP112 + VIM2 as described above

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407779
Original-Revision: 87e59ae1190c7db91e2110c6023163c21bff0775
GitOrigin-RevId: 6995705b60ee1b2ae417d117cb97f085f84d0a88
Change-Id: Iadfa0b4e3de0ce5ad97890f7b30bf810df1a2c47