[roll] Roll fuchsia [camera] Consolidate zircon.device.camera with ddk.protocol.camera
This CL resolves an old issue. Originally, the redefinition of the same
struct in both FIDL and Banjo interfaces was avoided in favor of using a
C wrapper. This solution is imperfect, e.g. in the instance that hlcpp bindings are used, it will need to be manually serialized anyway. The ISP implementation currently works as is.

This CL also helps pave the way to enabling a FrameRateInfo struct which
is necessary for the low light feature.

Test: n/a
Original-Fixed: 37654

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405786
Original-Revision: e53422e671f0efbb70c424a7c8ce12a2ead3f6c8
GitOrigin-RevId: c9be36daa769dfaa24553f48f1f3861382af11ff
Change-Id: I2a187e46d12a6305c6132a3d4d545b7d0ee8bee1