[roll] Roll fuchsia [wlan-hw-sim] Use WLAN Policy updates to detect failed association
The legacy policy::connect() command returns after either [connection is
made successfully] or [timeout after 20 seconds]. Since the connection
will never be made successfully in this test, the command always returns
after 20 seconds. The wlancfg module will continue attempting to scan
and join saved networks in the background.

After those 20 seconds, the test checks the status of the interface to
see if the connection was successful. However, the interface may be
either in SCANNING or JOINING state at this time, because of the
previously described auto-join behavior.

This CL modifies the test to use the modern Policy APIs, which gets
rid of this race condition

Original-Bug: 56104
Test: fx test connect-with-failed-association
Multiply: connect-with-failed-association
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407855
Original-Revision: df4d09bf9664cefba8982ffa72df9245f2a449f3
GitOrigin-RevId: 445059bd6b15fd2c2d0e987d836d9d57a87cab42
Change-Id: Ifc4f1bb032bac2117fc3eb8a7257ae54bbd4e940