[roll] Roll fuchsia [brcmfmac] Handle failed calls to getentropy
Rather than asserting when a call to getentropy() fails, treat it
as a runtime error and allow it to be handled with normal failure
logic.

Entropy is used to generate a random mac address, which
happens in two instances: device initialization (for our initial
mac address when none is available from the bootloader), and for
an active scan.

To accommodate testing, this CL includes a BRCMF_CALL construct
that allows us to override calls to external libraries in tests.

It also includes a fix to the simulator to properly cleanup when
initialization fails (which wasn't a problem in any of our
previous sim tests).

Original-Bug: 51987
Test: Included (bad_entropy_test.cc). Also ran manual smoke tests
      (active scan, passive scan connect, ping, disconnect) to
      verify that basic operation was not affected.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/408954
Original-Revision: 14eafa3690bc4038d189a9d16aca72828b06e2b8
GitOrigin-RevId: 2395cabb23aa989087d77d48e741526e29e61469
Change-Id: I88dc3283596b1ccb271e216a8220ea0cead29213