[roll] Roll fuchsia [rust][fasync] Prepare for the removal of fasync::spawn*
TLDR: fasync::spawn functions are going away in preference of the new
Task::spawn API's. This CL prepares the way by replacing usage of the old
API's with usage of the new one.

We're deprecating and replacing the old spawn functions (spawn, spawn_local,
spawn_blocking) and replacing them with an API centered around Task objects.

Tasks own the execution of the underlying future, and if dropped will cancel
that execution. To regain the old, unowned execution that spawn API's offered
a detach method is offered.

This change series sweeps through the codebase and replaces spawn() with
Task::spawn().detach() and spawn_local() with Task::local().detach() in
preparation for the removal of the original API's.

It's hoped that code maintainers will then take the opportunity (when it makes
sense for them to do so) to move to a model where Tasks are not detached but
owned for as long as they need to be running.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409136
Original-Revision: b001eca8bbb3c4e95374a40c1209f5f55f6fc41c
GitOrigin-RevId: ea821176cf9c6d06b4a4168ce78756657c274456
Change-Id: Id665e6b403226a9a584aaeb0891070c93d3084fb