[roll] Roll fuchsia [ping3] Fake time during SequenceStore tests
A SequenceStore#give test was confusing response times with latencies.
This change corrects that by testing the calculation of latencies. To do
this reliably, an executor with fake time is used for testing.

SequenceStore#give returns a latency in the form of a Duration.
Asserting equality of the two latencies doesn't make sense since they
should actually both be the same in the original test. However the test
wasn't failing 99% of the time since the test runner usually takes
longer than 1ns to reach the next call to SequenceStore#give.

Original-Fixed: 55572

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/408756
Original-Revision: 13a85a9da35f76eecce6866b398c255b22446a87
GitOrigin-RevId: acc23915ddbe5187d3fc9e3b87cf1d2ebe8878d8
Change-Id: I2d8a5b3cb15e950b4b69ac21f65d278b4402232a