[roll] Roll fuchsia [fidl][llcpp] Adds ReplyWithStatus() to Completer.
This CL adds a ReplyWithStatus() API, primarily for the purpose of
debugging FIDL connection errors.

On any failure, Reply() will automatically Close() the channel, causing
an epitaph to be sent with the status from the failed internal
operation. This is problematic as it may overlap with any contract
between the server and client regarding epitaph values and their
meanings. Additionally, the server-side user should have the option NOT
to close the channel depending on the error.

This change additionally adds a new internal API InternalError() for the
purposes of propagating information on the cause of an internal error to
the OnUnboundFn hook. Unlike Close(), InternalError() does not send an
epitaph, enabling the server or client code to decide how best to
handle the error, e.g. by sending a protocol-specific epitaph.
UnboundReason as well as the accompanying zx_status_t given to the
OnUnboundFn have been merged into a struct UnbindInfo with reason (enum)
and status fields. The kInternalError reason has been expanded into more
specific reasons, e.g. kEncodeError.

Original-Bug: 42086
Test: fx test fidl-walker-tests (previously fidl-test and
fidl-async-test-test)

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391627
Original-Revision: 224880ea2d26e4347b60d0a0342581e9f4c96c44
GitOrigin-RevId: a17b0ada4ec009c1d7ecc09d50aba7bf012cdf65
Change-Id: I61238f229f6a5f65f681bc1d3dc885c57dfaa19f