[roll] Roll cobalt Update comment on IntegerObservation proto
Previously this comment said that an IntegerObservation
could contain either a 0 or 1 value when generated for
a UniqueDeviceCounts report. However, we will never send
a value of 0 for an event vector that did not occur:
instead, we'll omit that event vector from the |values|
field of the IntegerObservation.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt/+/408759
Original-Revision: 494388b3cd91fcadb00ca62a715f346d08fd23ec
GitOrigin-RevId: 869020a08aa1afbf3cef4ebf9b78a1d881ae260a
Change-Id: Ia6c10923feca5d0f8c8a5a4a2453b0bcd71a86c3