[roll] Roll fuchsia [scrutiny] Package added to data model.
This change adds the package data to the data model which provides
useful information about which packages contain or use which shared
objects. This is useful for auditing packages as it allows the user to
quickly determine the scope of a vulnerability in a shared library by
cross referencing it with the packages it is in. I plan to extend this
in a future CL to enable parsing of the ZBI to extract the
component_manager manifest.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409320
Original-Revision: be938867ae0d376c51bff223e43da8167d7642ca
GitOrigin-RevId: 279fc9e10295db20f316764239591857a6e14fd7
Change-Id: I387f1316f132e576696e2b068825c25a8fc3dffc