[roll] Roll fuchsia [camera] Report aspect ratio in controller configs.
This change updates the stream image formats reported by the controller
to include the correct pixel aspect ratio. This allows a client to
display the content with the correct ratio, or select a crop region that
restores a unit ratio if desired.

This change also updates camera-gym to scale scenic nodes according to
the aspect ratio so they are displayed at the correct ratio.

Test: Ran camera-gym and observed corrected display aspect ratio.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409146
Original-Revision: 6b58037f9a723292c51e2bffc8a50ee835b6ec0b
GitOrigin-RevId: da41f93272c6b6febb6ccd102823c7a8c6168387
Change-Id: If8bfb8597f03c3674bd1a9b0088b82abefda1486