[roll] Roll fuchsia [scrutiny] Package search.
This adds a simple web view to allow you to search through all available
packages in the system for particular files. This is nice if you want to
search for a vulnerable library and see all the packages that contain
it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409701
Original-Revision: 4cf8deabdef511cc106a7f29de31fa8cbc34c602
GitOrigin-RevId: 37e24ab431e234ac2aae2a0da876a9f6d8f091a2
Change-Id: Id35ef64b479dba5ae6943a798c9f19a950fbf027