[roll] Roll fuchsia [shutdown-shim] return mexec call when asked to terminate
The mexec shutdown operation is different than the others, in that
driver_manager does not change the system power state during component
shutdown. It merely shuts down drivers and exits, after which the client
which requested the mexec will perform the mexec syscall.

When driving shutdown manually, the shim should notify clients when
mexec work is complete and the client may make the syscall. The shim
does this by initiating a component shutdown, and finishing the mexec
call when the framework asks the shim to exit, as this means that
driver_manager shutdown is complete.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407960
Original-Revision: 6c84a79dafa4f2147c6d8bbce99846d4c2c3258e
GitOrigin-RevId: f5194c66ada7a4f6e9de6a151f6c48e255065a75
Change-Id: I501254e2b373192bfce1b040933cddc18caf8f8e