[roll] Roll fuchsia [optee] Unpin shared memory pmt prior to bti close
The OP-TEE driver was letting its BTI handle drop immmediately after
pinning the shared memory region. This CL extends the lifetime of the
BTI handle to the lifetime of the device. The PMT is held by the
SharedMemoryManager as part of a ddk::MmioPinnedBuffer. It will be
unpinned when the SharedMemoryManager is destroyed. This CL also
implements the DdkSuspend hook for the controller so that all active
trusted app sessions are terminated and the shared memory is unpinned.

Original-Fixed: 56233
Test: fx test optee-unittest
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407213
Original-Revision: d24ef6ad0eab8eb66123384a209d679b5584cb39
GitOrigin-RevId: 739ac826b3a363d5022806226dd6c31553430ef5
Change-Id: I14416f2834082ca4c6283d2a9995bcb0fe402b01