[roll] Roll fuchsia [e2e] Also test for time zone *changes*
The dart runner should respond to time zone changes, as well as
time zone initialization.

This change modifies the timezone e2e test to try reading the timezone,
then change the timezone again, and ensure that the clocks match up
inside the dart VM and outside.

Previous test tried to do this only once, and didn't test repeated
attempts.  This works for the "deprecatedtimezone" implementation,
because it does not consult the time zone when computing the time
offset. But the "intl" implementation is sensitive to this, since it
asks about the timezone to compute its offset.

Original-Bug: 42387
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409840
Original-Revision: bb8163c2020c6f0490fd1a58526de6badf5d9f7a
GitOrigin-RevId: 4613142c1b727997992cf73a13e443be585e9589
Change-Id: I1b23bc1284228ff95f9c074a6e3b432c13c24432