[roll] Roll fuchsia [kernel][pc] Fix spinlock contention in UART IRQ handler.
Fix over synchronization between TX and IRQ paths that causes the
IRQ handler to spin for long periods of time on a spinlock when
there is a lot of TX activity. Because IRQs are handled on CPU 0
this can cause tasks that are assigned to CPU 0 to be significantly
delayed by the time "stolen" by the spinning IRQ handler.

Synchronization within the IRQ handler is unnecessary because the
TX, RX, and IRQ paths do not perform any read-modify-write ops on
UART registers and the TX and RX functions are independent. Also,
the cbuf and event objects shared between the IRQ handler and TX
and RX operations have internal spinlocks to protect invariants.

Original-Fixed: 55085
Test: Observed traces of continuous TX activity, written from both
thread and IRQ context. Spinning is observed before this change but
not after.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409519
Original-Revision: 2d287badcd13236a1bdd5694b642801564e63a4e
GitOrigin-RevId: 5663399b57202ea9e41063d82d9abb9a40d95a5e
Change-Id: I23ed26f77dfc1268d049a83b58ca4724dfdd66d7