[roll] Roll fuchsia [camera] Hide camera-gym views when they have no content.
A recent change updated camera-gym views to darken, rather than hide,
views when in the muted state. This caused nodes with no content to
appear as solid gray. This change restores the previous behavior so that
when the nodes have no content, they are invisible. When they do have
content but are muted, they are darkened.

Test: Ran camera-gym and observed expected behavior.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410055
Original-Revision: 236ef891dafc196ceee96ac4de15b346f0948245
GitOrigin-RevId: 4d1c9b08aa99604356af372df5a441ac9ced4cab
Change-Id: I7e7176bec17547dd7f6b8b27d2e74bcf339da4df