[roll] Roll fuchsia [servicefs] Allow passthrough of flags intended for remote dirs
ServiceFS checks flags on every Open request irrespective of
whether the path goes through a remote directory.

For example, when attempting to create a file in a remote
directory (such as a filesystem), ServiceFS will throw an
error even though the CREATE flag was intended for the
remote directory and not for ServiceFS itself.

To solve this, now ServiceFS first descends through the path and
then checks the flags only if it is opening a local directory entry.

Test: `fx test fuchsia-component-tests`
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409928
Original-Revision: d9bffa8582e0523b0dab0353dee522eee93e3bf9
GitOrigin-RevId: a019b6ae92dac04b36a7b8951be244809779a4dc
Change-Id: I0428b1ba0f7c08ee8aea57093ef7cf6eaab964ad