[roll] Roll fuchsia [i18n] Allows lenient localization.
Originally, the l10n JSON files generator errored out if it noticed
an original string that has no translated equivalent.  The intention
there was to point out that translations were missing.

However, this does not work in daily development, where we must have a
temporary discrepancy between the original and translated files, while
we wait for translations to be delivered.

This change introduces a flag `--substitute-original` which substitutes
"UNTRANSLATED(foo)" for message "foo" if a translation is not available.
The idea is to offer a clear way to know that there are faulty strings.
In this case, the expectation is that we could find them in QA.

It is always possible to re-run the tool in a checking mode without this
flag.  There could be other modes, e.g. omit missing string, but we will
not add those ahead of time.

Original-Fixed: 56958
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410100
Original-Revision: b465657cfa244804c0f7ca00e9ba7fa7bd2a09b5
GitOrigin-RevId: bfacb8002378ab08c5d6545c27958875678c8fc5
Change-Id: I532023a42ee0cbcd2da5d1625e008d382721a05c