[roll] Roll fuchsia [fidl][llcpp] fix ResultOf
The buffer in ResultOf is only valid if the call/write was successful
and, in case of call, if the decoding was successful.

That means that the buffer is valid only if |ok()|.

We have to release handles only if the buffer is ok.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410124
Original-Revision: ad483d72bd276ed63d3eaa9b7848ea34a220cdaf
GitOrigin-RevId: 6f9bb9f2a88681b5b14d9cf7c2d87147eee0c38b
Change-Id: I11f89dab56d3470cc165f3cbcb5247f0b5f0e4e5