[roll] Roll experiences [benchmarks] Add flutter-driver option to scroll-flutter
Previously, flutter_driver support was added using
flutter_driver_extendable in BUILD.gn. However, this is only effective
on debug builds of Fuchsia. We want to be able to use this for automated
tests in release builds, so replace it with an explicit command-line
argument that enables flutter_driver in the scroll-flutter app and set
product = false in BUILD.gn to force a non-release Flutter build.

Original-Bug: 55832

Test: Manually tested in FEMU with a WIP benchmark; this appears to
allow it to use flutter_driver even on release builds of Fuchsia, where
previously it was only usable on debug builds. Removing the
--flutter-driver argument when launching the app causes the attempt to
use flutter_driver to fail in the same way the previous version did on
release builds.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/409634
Original-Revision: 77dacb00187497d10d17fcffa91aa6027cfa5b3a
GitOrigin-RevId: 2999e1e93d18671466ea10186be49d14e2f2ecb5
Change-Id: I2919c302726b3e600046190688836c223ab64021