[roll] Roll fuchsia [GN SDK] Handling gsutil missing in pave and serve
This changes the fuchsia-common functions for running gsutil
and cipd to return a non-zero code when the commands cannot be
found instead of exiting.

This is in response to a developer report of fpave.sh just returning
without an error or success.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410194
Original-Revision: d19bdebf3b7eb858e531d4b7a65d727b0d076a53
GitOrigin-RevId: d0d3a522641bea2dcab23c0170a84cac829e6dab
Change-Id: I7ef0ba7c37669d4f56795f510273dabe94232b9d