[roll] Roll fuchsia [fidl][go] Allow unused reflect import in conformance tests
When certain GIDL tests are commented out, this import is unused which
means that you need to go into the go geneartor and temporarily comment
out the import. In order to avoid this, always have a dependency on the
reflect package.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410048
Original-Revision: 8ad040efad4358f0dd9bce6826fde55bbca1d2c3
GitOrigin-RevId: cadad205cd8353b8e41d1d96b95785a2ea982f3f
Change-Id: Ib4eef2ae1ab407c7e6635a3b8e0ca0c1ef27d419