[roll] Roll fuchsia [fidl][llcpp] Reply not required after unbind.
For messages which require responses, CompleterBase currently asserts
that a reply has been sent within its destructor. However, this doesn't
take into account the possibility that the channel has been unbound, in
which case a Reply() or Close() on the completer would be a wasted
operation. This change adds an API to Transaction which enables
CompleterBase to determine whether the binding is still active. By
default, this API returns false to match the existing behavior in other
implementations of Transaction.

Original-Bug: None
Test: fx test fidl-walker-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410217
Original-Revision: 6b6ebc9f3302de4beed628742fe760785a6ff3cb
GitOrigin-RevId: d59d0acf2e61be59c67235a2c1c645f9e5fadc98
Change-Id: I08e0d6c2a46240453ace223691351df9875ec642