[roll] Roll fuchsia [triage] Short circuit boolean operation evaluation
This CL ensures that we don't attempt to evaluate metrics in boolean
operation after one metric would determine the result (a single False in
an And or a single True in an Or).

This can be used to ensure that missing metrics are suppressed if an
earlier defined metric in the boolean operation sets the result.

This also adds unit tests for various metric operations, including the
new behavior.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/402133
Original-Revision: fb716dcac6df5d2986eb1c6d318a90b63a794b0a
GitOrigin-RevId: c1e9ec2009c3f581ce3a9623895f42f5d0722f2d
Change-Id: Id9780ba53b7e352000ef7ef84f4482569204fbc6