[roll] Roll fuchsia [kernel][iommu] Improve comments for whole_segment field
IOMMU descriptors have a list of scopes (devices) associated with
each segment. whole_segment describes whether the list of scopes
is the list of devices managed by the IOMMU or the list of devices
not managed by the IOMMU.

Improve the comments to make that clearer and to align with the
respectful code guidelines.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410197
Original-Revision: 142d7826826b7f24a4a3794af89b1c55b2a3e078
GitOrigin-RevId: 7c4a2b39db56b2540b88e67606d0d1986010a010
Change-Id: I5a816111388cc8568e0e45fe12eebfce5b8d8f0e