[roll] Roll fuchsia [diag_tool][diagnostics] Support iquery output in diag_tool
The output from iquery is missing some fields expected in the
diagnostics schema, which prevents it from parsing. This change modifies
the schemas to replace those fields with defaults if they are not
present so parsing completes successfully.

The iquery output may also be unsorted, so this change sorts all
hierarchies before diffing to prevent noise in the diff results.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409206
Original-Revision: e0c07e794201dac86833e2566c5ad0088f212025
GitOrigin-RevId: bfca2e62eb9cfbab60116fccedbe880333c78cac
Change-Id: I61181c52ac22c93b4c35e241c1ab7cc94e11e9f4