[roll] Roll fuchsia [archivist] Support disabling pipeline filtering using a special file.
This will allow us to parse and verify selectors in settings where they
should not actually be applied. Inspect output for the Archivist
includes information on whether filtering is enabled or not.

Filtering is enabled by default, and is disabled by placing a specially
named file along with the config files.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409207
Original-Revision: cde4845284ddc96809a6bc5e206969c442a67aa5
GitOrigin-RevId: 5f031872434ba1f1db16302c753b6778261f323c
Change-Id: Ia80cd4fce4f5c8c74c24be575a161db275a2f494