[roll] Roll fuchsia [ffx] daemon start starts a daemon
The pid must be waited on unix hosts, otherwise the child process
becomes a zombie. This implementation will perform a daemonize
downstream of the fork, so it will succeed if the second fork works,
but will wait for the first fork to complete, avoiding the zombie
condition. This also prevents the hosting session leader from
terminating the daemon at some later point. The allowed deprecation
is for libc::daemonize that is deprecated as of macOS 10.5. We'll
need to bring in a 3p lib or write more ffi code to replace this
eventually.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409054
Original-Revision: 14d65bbef6d53a9bd4a38252d5f3867e38d2f35f
GitOrigin-RevId: 07a1c08c9756cd2f0e73be98a98fef2b7d9189b8
Change-Id: I1f562f8db39cd9c14a8527be2b87624a3b77e65a