[roll] Roll fuchsia [debugger] Remove ServiceDirectory from DebugAgent
The DebugAgent should not be using the system FIDL interface for
ServiceDirectory but functionality should be instead exposed by the
SystemInterface object.

Move handler querying out of DebuggedProcess and into ProcessHandle.
This is the last significant system call in the DebuggedProcess which
means it can be completely mocked.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410262
Original-Revision: 699b20c34d4ecbcd89d67beb3113d051af4b4aca
GitOrigin-RevId: 3d97aef95c3f44401238e9616a2ec1e893aaaaf6
Change-Id: I603cca770c587156a1c57f5b02002cbb12788875