[roll] Roll fuchsia [third_party][fatfs] Use 0 for root dir in .. entries.
Fat expects all '..' entries which reference the root dir
to have a cluster number of zero, even though on Fat32 this is not
necessarily the case. Fix this by checking if the parent cluster is the
root cluster when creating a directory, and setting cluster to zero if
so.

Original-Bug: 53796
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410160
Original-Revision: 0fd0964b008c668b602672727c6eae4d99f1e064
GitOrigin-RevId: 2df2476acda807535cfacea6fa4a72941da0a425
Change-Id: I0832e9b865329a558954e5773e49be094f5b0077