[roll] Roll fuchsia [ldsvc] New C++ fuchsia.ldsvc.Loader implementation
This is a rewrite of //zircon/system/ulib/loader-service in C++ and
using the LLCPP FIDL bindings, plus all the latest C++ goodies like
zx::status. The plan is to migrate all remaining usages of
ulib/loader-service, which is in C and lacks comprehensive test
coverage, over to this.

Original-Bug: 55959
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407649
Original-Revision: 727bba20ba0324970420af19f9c93a905df42275
GitOrigin-RevId: bdbd89987878e898eb8e4d1af4be78e0c676a3a1
Change-Id: Ia519d1c590aee862ee4c205e8e74951267450699