[roll] Roll fuchsia [amlogic_decoder] decode flake repro test
Plumb a test hooks struct.  On astro this demonstrates occasional
incorrect decode of the last frame in the sequence, and when the
decode is incorrect, it's always the same slightly bad pixels on
astro.  On sherlock it's not always the same bad pixels, and the
repro rate is lower.

BUG: 13483
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410220
Original-Revision: 606463a8c53156fff8bb2b181611a8b5426d95df
GitOrigin-RevId: 088ea50b53697cb69f75fed5ba74336aec3fa2f6
Change-Id: I4c4388b48b4e34384c79db86a4124d7509295c6e