[roll] Roll fuchsia [fidl] Move abi_tests to internal.h
These tests were not easy to find and make it harder to confirm that we
are verifying that the size doesn't change. In addition, the size check
was only happening when abi_tests is included in fx args.

By moving these checks to internal.h, they are always checked and easier
to find.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410274
Original-Revision: 683fa870ba17b5c281855ecaf419e0aa71c37728
GitOrigin-RevId: 1dfc6a91bb89094c7d24d42c2abe59d746c0ed7a
Change-Id: I42ba5e7aacdc78b1b342a75dbe3a327dc2240d0d