[roll] Roll fuchsia [dart_tool] conditionally generate a library
The dart tool does not need to generate a library unless it
defines sources. This will check if there are sources and
generate a library if needed. This allows us to depend on
multiple dart_tools that exist in the same library from
a single target.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410896
Original-Revision: 8ad10c6ad0fb205dac81c77c823b111435cd96fe
GitOrigin-RevId: 31c350c04db24bafd3559810e36eff4b13c50ad5
Change-Id: Id18abe82d9fa8375934f652792459124c4647806