[roll] Roll fuchsia [camera] Temporarily Disable DeviceEnumerationTest for Camera Devices
This CL removes devices from being run in DeviceEnumerationTest. This is
a temporary change to enable the soft transition of camera drivers to
ddk.protocol.camera.sensor.

WARNING: This is a breaking change for camera functionality and will be undone in a follow-up CL.

Test: n/a

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410781
Original-Revision: 5eced63bc9b19d28ee173ab603e2965111bbc628
GitOrigin-RevId: 48a1536dfda0109fedc89c520578cb67d47d523f
Change-Id: Ie9c385fd2875b85d126e0db60a40b472b7248287