[roll] Roll fuchsia [modular] Add session launcher components as a concept
A session launcher component is a component launched by basemgr that is
an entry-point to a session. If no session launcher component is
configured, basemgr will directly launch sessionmgr.

The session component is given access to the the
fuchsia.modular.session.Launcher protocol that it can use to instruct
basemgr to launch sessionmgr with a particular configuration.

Original-Bug: 54048
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407038
Original-Revision: 8992ce12d7fcdec51d358adc6983e772de2c149b
GitOrigin-RevId: d7c685b77c91996a6a4855b13925f2db33dac341
Change-Id: Ib640624d96e7464710f82636eb2c8d0d015091c8