[roll] Roll fuchsia [backtrace] Simplify backtrace request output.
The backtrace request library triggers an exception that is specially
handled to resume. But most people don't know this and the output looks
almost identical to a crash, is reported as an exception and a "sw
breakpoint". This results in confusion, especially when used as a
watchdog (Scenic does this).

This removes the output from the crash service that it's received an
exception in the bcase of a backtrace request.

In the inspect output, this clarifies that the result is a backtrace
request and omits the CPU registers and stack binary dump. Normally in a
backtrace request, only the actual stack frames are needed and this
output makes it more difficult to follow and look more like a crash.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410821
Original-Revision: d1c4677b58706d11eee0119011e0e531e0467c58
GitOrigin-RevId: 928e07dce0eef749403f2de5e0f08273077c0d19
Change-Id: I7d5a23e683feb6cb32d7cf45b008a145cb075018