[roll] Roll fuchsia [fvm] Migrate off DEVICE_ADD_INVISIBLE, DdkRemoveDeprecated.
The DdkInit hook replaces DEVICE_ADD_INVISIBLE. If implemented,
it is called after DdkAdd(), and it is guaranteed the device
will be invisible and not unbound until InitTxn.Reply().
Replying with an error will schedule the device to be unbound.

DdkAsyncRemove() replaces DdkRemoveDeprecated(). This schedules
the unbind hook to be called (and eventually the release hook)
and is safe to call multiple times.

---------------
DDK lifecycle:
---------------
1) DdkAdd()
2) DdkInit()
3) InitTxn.Reply() // Device becomes visible
4) DdkUnbind()     // started by InitTxn.Reply(!ZX_OK) / DdkAsyncRemove
5) UnbindTxn.Reply()
6) DdkRelease()

TEST=  fx test fvm-driver-integration-test

BUG=37831
BUG=43261

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407130
Original-Revision: 4709f5019d3c22ac741d661dd97781f3f54d0788
GitOrigin-RevId: fa148a435e5ce890ee7fe599f2c1c25cc2922e4f
Change-Id: If7a7f912377ed581652a88cee5d12ea1fdb6254c