[roll] Roll fuchsia [block-verity] Implement CloseAndGenerateSeal
Upon request, the driver will now generate integrity data, write it out
to disk, update the superblock with the device configuration and root
integrity hash, and return the hash of the superblock.

Added tests that verify the contents of the generated integrity data,
superblock, and seal against the values expected given the design
specification.

Added additional tests that verify that on any I/O failure, the
computation is marked as failed and completed and propagates the error
in question.

Thanks to Allison Pearce for tracking down why the integration tests
were failing in work-in-progress versions of this patchset.  She
verified that the driver implementation was correct, determined that the
test failures were due to misuse of FIDL (stack use-after-free) in
block-verity-test.cc, and supplied a correct implementation.

Original-Fixed: 55710
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409817
Original-Revision: 35d9ba361eacbd0b409b7fa34a1ba95ef58d9430
GitOrigin-RevId: 9ba388c002aafc94cd3bac42677076a5dca6ceff
Change-Id: I85f23a6414f068a2b57fe1c453b20997286f183a