[roll] Roll fuchsia [ffx] Reland, daemon start starts a daemon
The pid must be waited on unix hosts, otherwise the child process
becomes a zombie. This implementation will perform a daemonize
downstream of the fork, so it will succeed if the second fork works,
but will wait for the first fork to complete, avoiding the zombie
condition. This also prevents the hosting session leader from
terminating the daemon at some later point. The allowed deprecation
is for libc::daemonize that is deprecated as of macOS 10.5. We'll
need to bring in a 3p lib or write more ffi code to replace this
eventually.

The daemon does not chdir("/"), as it does not yet have context
information passed from the frontend sufficient to otherwise discover
the correct environment and configuration.

Original-Bug: 57054
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410452
Original-Revision: 7f785b0e5a69eea01862a0cdce8cc03bf404ae10
GitOrigin-RevId: 057ac66023ec5c1b75c1da7272f6c8bd7393880d
Change-Id: I465406a9a848563cc598f2a02cba1e4b97cd7ed5