[roll] Roll fuchsia [blobfs] Pager watchdog
Creates a new watchdog thread which logs any page faults that take
longer than a configured deadline (currently 60s) to satisfy.

When the deadline is exceeded, the watchdog takes a backtrace of all
blobfs threads, and logs the error. In the future we can extend this in
other ways if desired. For now, the watchdog never affects the behavior
of the pager thread, or of page fault handling in general.

Since the blobfs pager being stuck can effectively stall the entire
system, this is an important state to watch for, and will aid if any
bugs result in pager hangs (either bugs in blobfs itself, or cycles
where a process that blobfs' pager thread depends on takes a fault).

Tested: Manually (set small deadline to verify logging, set real
        deadline to verify not-logging); new unit tests.
Original-Bug: 56818
Original-Bug: 56934 # Fixed
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410050
Original-Revision: d32cafd2c5dcbd3a50cf7eb7b2d1a17982358442
GitOrigin-RevId: e4c5d51f406fa0455b1b003a80b3866bf5e0251c
Change-Id: I7026842d3022eeaa6120de2df025988d49fa5fcb