[roll] Roll fuchsia [camera] Relax asserts on InterfacePtr::Bind failures.
This change relaxes existing asserts on the return value of InterfacePtr
::Bind calls. Specifically, it allows for ZX_ERR_CANCELED, which can
occur if the default dispatcher for a thread is owned by a loop that is
being shut down. This can occur if an object is destroyed before it is
fully initialized.

Test: Reproduced by rapidly changing the configuration in camera-gym.
      Verified that with fix applied, assert no longer fires.
      Added DeviceTest.BindFailureOk test.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410288
Original-Revision: 8fc6473fd229d035295b56717800cea2e6935cd2
GitOrigin-RevId: 0a9145a212b85d65509c8059b04698cb762a7908
Change-Id: Ie354374ece32f77d98cbad8c6252e9f620575155