[roll] Roll fuchsia [camera] Fix binding leak on pipeline creation failure.
This change fixes a bug in the controller where a pipeline that fails
part-way through creation left the ISP bound to a non-existent node.
This can occur if sysmem is under pressure and an inter-node collection
fails allocation. The change adds a shutdown auto-call, and a new check
on shutdown to only propagate the shutdown call if a child node exists.

Test: Reproduced binding leak by shortening camera-gym cycle time.
      Verified fixed by rapidly cycling configs 100 times, then cycling
        as normal.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410264
Original-Revision: 08808399ce73903bddef1e43a2c0b22485779b3f
GitOrigin-RevId: 3c760a2eee1933497012af599988c8dfd7eb9792
Change-Id: I0b0cbcf7f7f059068d82e774086a021c50a850f8