[roll] Roll fuchsia [testsharder][testrunner] Record results from all test tries.
Testrunner currently allows retries but only records the result of the
last attempt. This hides flakes, so we should report all test runs and
let the recipes deal with how to handle flakes vs failures.

Testsharder is also updated to use a run_algorithm instead of
max_attempts to determine how to use the `runs` field when running a
test. For example, the run_algorithm will either tell testrunner to keep
running the test `runs` amount of times or to attempt the test up to
`runs` amount of times until it gets a success.

Original-Bug: 56702
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409059
Original-Revision: 3054aec908506ace765ed5b0a0d6a388b38ea236
GitOrigin-RevId: c7f12fb9e9626cf4bf3f94b9ae60307db1025771
Change-Id: I96821242ccaff184aeb5a1f720de44495922b4eb