[roll] Roll fuchsia [fatfs] Fix directory rename issue.
When renaming directories, the ".." entry needs to be updated.

This fixes:

  PersistRenameLoopTest.MultipleRenamesCorrectAfterRemount/
     FatfsRenameDirectory100TimesThrough15Directories

Whilst addressing this, I noticed that the implementation writes
long-file-name entries for the '.' and '..' entries, which might run
afoul of some fsck programs (although not ours), so I fixed that.

Original-Bug: 53796
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/411395
Original-Revision: f2dc5b2220c8701cc7eb514556e96e767b525251
GitOrigin-RevId: 97259641da5ce57bdd4004ddcf503531396ae483
Change-Id: I0fbc4a4efba179fcbae67a4fd747b9084613398f