[roll] Roll fuchsia [usb][hub] validate descriptor returned bLength
The hub driver requests a number of descriptors from the hub device.
The current implementation is not verifying that the length, as reported
by the descriptors, equals the total returned descriptor bytes.  If this
happens, it might be a malformed descriptor which has the potential to
leak out-of-bounds memory.

This adds .bLength (and .bDescLength) bounds checking.

Original-Bug: 57060
Original-Bug: 57059
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410903
Original-Revision: 7490009c94ed6677cfbecf6e22f1abc1c797f4b5
GitOrigin-RevId: 80a6809881e5510ec26acce5fca8ad3c44c9ab70
Change-Id: Id0183b7f6399e5f85cfb7bf5c9bbed2e3f3604aa