[roll] Roll fuchsia [debugger] Improve suspending.
Clarifies the two types of suspensions. Previously there were several
different ways to suspend and resume a thread, and there was also a flag
to indicate whether the client thought that the thread was suspended.

This change separates "client" suspensions that the zxdb frontend
considers its domain, from "internal" suspensions which happen for
things like stepping over breakpoints. Client suspensions have a simple
on/off state, while internal suspensions use suspension tokens.

Creates a SuspendHandle wrapper for suspension tokens so that these can
be used in mocks.

Removes the existing "ref counted suspend" machinery which emulates part
of this in a complex way. Delete the unit tests for this.

Moves waiting for suspension to the ThreadHandle so this can be mocked.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/411275
Original-Revision: a0868527ed59bb8f46f2e8298a916f8c37136790
GitOrigin-RevId: 0da3cd1798dcc050988f72bf8aa94339043f12e1
Change-Id: I9b48af06d4be3328c225580223cd708cd3f5ebb9