[roll] Roll fuchsia [triage][codelab] Update monikers to match real ones
Current monikers in the bugreport and in the codelab do not match real
ones. A previous CL (https://fxrev.dev/407292) updated the solution
to use real monikers. This CL makes sure that the codelab uses them as
well.

An advantage of doing this instead of keeping a fake example is that
the codelab resulting triage file can be used in real-world bug reports.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/411560
Original-Revision: 1e7a6248d409c2174a976e3239b9db52ad73b524
GitOrigin-RevId: 3275569f39b1745e2846db4196ae928bdbd36fb7
Change-Id: Ic0769bb4ca910ab5645d847ffcd2bc4e0814de3a