[roll] Roll fuchsia [docs] Replacing "SDK" where we mean "IDK"
To make the separation of concerns between the "Core" SDK and the
"Frontend" SDKs, The "Core" SDK is now referred to as
Integrator Development Kit (IDK).

The IDK is the development environment independent artifacts that
define the public Fuchsia platform. This includes system libraries,
headers, FIDL files, and Fuchsia specific host tools needed to target
Fuchsia as a developer.

The IDK is the foundation of SDKs created by development environment
integrators which augment and transform the IDK to be usable in a
specific development environment.

This is being done now as we continue to refine the processes
and stuctures around delivering an end-developer ready SDK.

Original-Bug: 51916

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409263
Original-Revision: 1c6e2e4758ef1b8e7a016bb08bff1cfde697822b
GitOrigin-RevId: e0a07c6c8529340e9e2ac47294b662f4d1ba80e8
Change-Id: I24b1e7f9f5a905d3d5abf938c69e6c99a69bea0a