[roll] Roll fuchsia [cobalt] Convert the prober into an e2e test.
In this change, a new package is created to run the cobalt_testapp in prober
mode without using run-test-component. Since the prober requires network
access, it is expected to not be run as a test (the test component
framework recently stopped allowing tests to access the network services).
This new non-test package is then run by a new end-to-end test for the
prober. E2E tests are allowed to access network services.

Subsequent changes will update the infra to use the new end-to-end test in
place of the current test package for the cobalt prober CI, and then remove
the old prober package from fuchsia.

Original-Bug: 57076
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410943
Original-Revision: 67472cb351db697ece1eabd830c93e542cd55150
GitOrigin-RevId: 795f6a112a48c9afc49580ec4e6ecff487d87672
Change-Id: Ica3c348191e777af9734e95ad1730e535ae7d972