[roll] Roll fuchsia [cts] Add CTS build templates
These templates are used to create a CTS test. A cts_* GN target should
be able to depend on other cts_* targets and carefully selected non-CTS
targets. This is done to ensure that the CTS only uses / depends on
things that are released in the SDK.

In order to label a target as a CTS target, each cts_* GN template
creates a file called "${target_name}.this_is_cts" whose existence is
verified to ensure a target is a CTS dependency.

TEST:
Run the sample test with:
```
fx set core.x64 --with //cts
fx test //cts
```

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/398817
Original-Revision: 3c407c1b894e03a5a25ba6076e8627d2293ed542
GitOrigin-RevId: cd6e67221f7b9c4bc6787043576202df84c8f12c
Change-Id: I78120d135a0a919f147488b290810812477d0a18