[roll] Roll fuchsia [GN SDK] Improving fserve-remote
This clears the default IP address and sets the default device name on
the laptop running fserve-remote. Then on the remote desktop, the
default IP address is set, and the device name cleared.

This eliminates any stale default values which could conflict with the
remote serving.

This also adds the sleep invocation to keep the tunnel open vs. closing
quickly.

It also adds "-i" to the shebang for fserve.sh so it uses an interactive
shell, meaning the path and other environment variables are setup
correctly. This fixes the issues with not finding gsutil.

BUG: b/162249694
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410994
Original-Revision: 7987669458763d730d6b931f518901721d52a34f
GitOrigin-RevId: 5a5923cef22d474457071b3e0f1f090fee317109
Change-Id: I61c5574c49cbad39900643e742396ebc35a40f6d