[roll] Roll fuchsia [wlan] Return BadCredentials after too many EAPOL handshake timeouts
The LinkState state machine inside the SME state machine will return
EstablishRsnaFailure::KeyFrameExchangeTimeout if too many timeouts
occur when exchanging key frames during the EAPOL handshake for RSNA
establishment. Assuming all cryptographic parameters associated with
the handshake are read and used correctly, it is most likely the case
that bad credentials were provided by the client. Specifically, in
IEEE 802.11-206 Section 12.7.6, any time a handshake key frame contains
an invalid MIC, the receiver of the key frame silently drops the
frame. When a KeyFrameExchangeTimeout, wlanstack now returns
ConnectResultCode::BadCredentials from SME.

This change also reduces the severity of the message logged while
the authenticator processes the second EAPOL message. It is now
a warning instead of an error. A warning fits better because the
authenticator does not give up in this case. It just awaits the second
message again.

Added the connect-with-bad-password hw-sim test which tests that the
policy layer sees the ConnectResultCode::BadCredentials from SME and
uses it to return a ClientNetworkState with ConnectionState::Failed
and DisconnectStatus::CredentialsFailed.

Test: fx test connect-with-bad-password
Original-Fixed: 53234
Multiply: connect-with-bad-password

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410474
Original-Revision: 9db263babc25c873b5e073af62a4584bab69f78f
GitOrigin-RevId: 6d464866f9dddfa0513dcbc8e5210ba4d096fcd1
Change-Id: I88d55b2beab6dbdf48488813907a542730da1592