[roll] Roll fuchsia [ffx] [docs] First round of a getting started doc.
This needs some filling out, especially the config and "multiple
devices" sections, but I think as-is it will be valuable for onboarding
people to `ffx`.

I decided not to commit too hard to making it tutorial-style. We may
want to revisit that later.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/411249
Original-Revision: 3f7f8f772428529f190dff77f1b31361ce3e52c4
GitOrigin-RevId: 38fe7e799594ebc4a00ff9dfdd23ab487acf7cbf
Change-Id: If7f17fd317d973ee7c8137f9744efd1d9747451b