[roll] Roll fuchsia [drm] Add KeySystem.DestroyDataStoreAndConfirm
When a data store is added to a KeySystem, it includes a
fuchsia.io.Directory that is used for storing data. When this data
store is destroyed, the KeySystem should clear the contents of this
directory and release all references to it. By sending a reply after
this is complete, the client can ensure that it has been properly
cleared.

Original-Bug: 13619
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412045
Original-Revision: 03aaf69a128edd92f7560a926b8168267283e438
GitOrigin-RevId: cf36457a94f5dde941cea82ee9b2ca2470d23dfc
Change-Id: I8f30c4c7182e904c89f541ed31cd6e2e5f9c3a7f