[roll] Roll fuchsia [tefmocheck] Add a check for QEMU segfaults
This is one of the more common unattributed flakes I've noticed. For
example: https://ci.chromium.org/b/8873278597020207616

I also reworked the way botanist logs these errors so it doesn't seem
like they're coming from testrunner. Specifically, we only log "command
[testrunner] failed: <err>" if the error that occurred is a subprocess
ExitError.

In the case of this failure mode, the log previously looked like:

  botanist ERROR: command [testrunner ...] failed: QEMU invocation error: signal: segmentation fault

but now it will look like:

  botanist ERROR: QEMU invocation error: signal: segmentation fault

This is better because the error is unrelated to the command being run,
so the old log was misleading. It also makes it easier to match this
string in tefmocheck, since there's no command string in the middle of
the log.

Original-Bug: 57463
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412499
Original-Revision: a584524141568c2de069edc5572bce67e26f7308
GitOrigin-RevId: 0c7d2ed4f1848d2ddbad9a93c9cee663d4418d12
Change-Id: If4e5a20550592b8e0c154a8c91f4ac5955fdd047