[roll] Roll fuchsia [power_manager] Remove mexec support
The power_manager's implementation for
fuchsia.hardware.power.statecontrol.Admin/Mexec isn't quite correct in
that it treats Mexec the same as any of the other supported shutdown
states. This assumption isn't correct because for Mexec, the shutdown
provider is not expected to actually shutdown the system.

Since there is no scenario where power_manager is expected to handle a
Mexec call (it is used in zedboot, where power_manager does not exist),
power_manager should just return NOT_SUPPORTED for this call instead of
maintaining or fixing this broken implementation.

Test: fx test power_manager_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412050
Original-Revision: f79185e6677653588552736029446b2e5bf0dcd9
GitOrigin-RevId: 7c530d5930b1a5199425f354cfbf7ee0fdbcab7a
Change-Id: I72e38a480d10499f12464ae88ff4d403dd936284