[roll] Roll fuchsia [Docs] RFC: Units of bytes
Why is this a priority for current work: This issue was discovered
because memory is at a premium. While debugging, it was found that
different tools gave values for byte counts in different notation.
These discrepancies are large enough to cause concern and confusion.
Being constant in our measurement notation is expected to streamline
this work.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/397239
Original-Revision: 79b6c18eb7526da4a81b332e649806fb3e215a0c
GitOrigin-RevId: f1f296ffa8548f2bec4349602ec08ba7dc6650ae
Change-Id: I22e23ffba0501fbaac04e5810b385285ba9f7aa9