[roll] Roll fuchsia [sshutil] Use Conn instead of Client for test
This test case previously contructed an entire sshutil.Client instead of
an sshutil.Conn. This is unnecessary because the Run() functionality
exercised by this test case is entirely within the Conn struct. In fact,
the Client.Run() method is just a wrapper around Conn.Run().

I suspect using a Client here may have been contributing to the
flakiness we've seen in this test case. The setUpClient() helper method
that the test case was using to create a Client actually starts the
underlying Conn's keepalive goroutine, which adds some unnecessary
parallelism and nondeterminism to this test. Even when the test passed
because Run() exited as expected, sometimes it was because a keepalive
failed and closed the Conn, and sometimes it was because the ssh request
to actually run the command failed. By switching to setUpConn() and not
starting the keepalive goroutine, we ensure that the only way Run() will
ail in the test is when the request to run the command fails.

So even if this doesn't completely fix the flakiness, it at least makes
the operations that happen during the test more deterministic, which is
always a good thing.

Original-Bug: 50198
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412656
Original-Revision: 7db39b7f93cf4806661a7a648790832ac302bbf1
GitOrigin-RevId: d3b8fe366a0385c2aeb68f20a4404d01ae7ef671
Change-Id: I1b62ed148c914336b67ee5eaaae5d9e0bb95afbe