[roll] Roll fuchsia [cmc] cmc accepts name or path for routing protocol and directory
As an intermediate step, allow cmc to accept use/offer/expose
protocol/directory to reference a capability name or a path.

This CL does not include cm_fidl or routing changes.

Note: This change adds a noticeable amount of complexity to handle both
names and paths. We expect this to be short lived, as the path support
will be deleted once we complete the transition.

Original-Bug: 56604
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410949
Original-Revision: 82dbfbf6426d3557815fc356cba265d50ccbb4c1
GitOrigin-RevId: f203c91d0cac0246cb21ef4270a7ce774a67bf62
Change-Id: I500a70e1b710a580245af6f199ab7a8d3d92f141