[roll] Roll fuchsia [zxdb] Handle new tombstone value
With https://reviews.llvm.org/D81784, -1 is a reserved value to indicate
a removed function. zxdb initially tried to use this as the starting
address for an address range.

This patch applies a fix similar to breakpad where we ignore entries
with a value of -1.

TEST=Checked locally to assert we don't see this error with the new
clang and llvm prebuilts.
Original-Bug: 56893

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412694
Original-Revision: d778aa00305f56af7f173b231b7c6a9d39865096
GitOrigin-RevId: d07a0238c9be1226a89de0b84ed998434fcffd06
Change-Id: I3a932e571cd36b1445c38d63a24f3c187bee5dbb