[roll] Roll fuchsia [camera][sensor] Add analog gain support
This CL adds support for programming the analog gain in the sensor.
To promote reuse of the code, details of the analog gain encoding
and limits are read from standard sensor registers.

As this is the first piece of exposure setting implementation,
it also introduces the deferral of register writes and their
synchronization through the grouped parameter hold mechanism.

In support of these additions, a 16-bit I2C write routine was added.

Note that this CL only adds support to the CameraSensor2 interface.
This is the interface currently used by the ISP, and the older
interface will soon be removed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412554
Original-Revision: 25788d43fa4e48d54e9588e91c511b6f7dccc909
GitOrigin-RevId: 9eb60dcc3dd56750187b8d2ef306fcbe1d21f36c
Change-Id: I98bfb19833191971367a5b33c34933b3f95c533f