[roll] Roll fuchsia [amlogic_decoder] Keep cores powered up for entire lifetime of decoders
The current code powers off a core when an instance is swapped out and
powers it on when the instance is swapped in. This causes a lot of short
power-off/power-on transitions, which seems to cause some rare hangs. A
workaround added to increase the time of those transitions slows down
decoding, so we should reduce those transitions as much as possible.

In practice, idle decoders remain scheduled and powered on, so the only
time this change will cause a core to stay on when it would otherwise be
powered off (for long periods of time) is when context switching between
VP9 and H264, where before the unused core would be powered off.

Original-Bug: 57380

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412400
Original-Revision: bb9e6db3a0f2880988ca3d8465a33c2d488073eb
GitOrigin-RevId: 746ac81d0697e3675fe4098d77b5a47b45bedd15
Change-Id: Ia31eb67b6d1c49a120dd2e3bcfae91424161550f