[roll] Roll fuchsia [lowpan_driver_common] Make `serve_to` method return future with `Send`
The `ServeTo` trait was being defined in such a way that omitted the
`Send` trait from being associated with the returned future. This
prevents the returned future from being run on a different thread.
This commit removes that limitation by ensuring that the `Send` trait
is properly associated with the returned future.

Test: fx test lowpan-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412709
Original-Revision: 19077807f2f21fda227eb801cc8923891359e8af
GitOrigin-RevId: 0ad417cb4654387326501a57e48e4aa1e8a57ba7
Change-Id: I305743c07e87c8aa994677006205ea1a8d2a2bb8