[roll] Roll fuchsia [camera] Fix new flake in Device test.
This change fixes a flake introduced in fxr/411922 related to stuck-
client buffer recycling. Previously, the test did not wait for a frame
to be received before sending subsequent frames. This caused the test to
erroneously pass most of the time, because it only verified that a
specific frame was eventually recycled, not recycled in a specific
order. Additionally, the test was using inconsistent values for min-
camping-buffers, which could lead to deadlock. This change fixes the
issue by waiting for each frame to be correctly received or recycled,
and by using a common constant value for min buffer count.

Multiply: fuchsia-pkg://fuchsia.com/camera_tests#meta/device_test.cmx: 100
Original-Fixed: 57256
Test: Ran DeviceTest.GetFrames with repeat=10000.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/413022
Original-Revision: afaee75fa7ee4418a17b3f0bf65b26aefddc3b50
GitOrigin-RevId: 7d69de598a7e709d6fc920c47656aa980b1f7016
Change-Id: Iadf2615ba37db32b34aa0ce4df5fe66d5678cf92