[roll] Roll fuchsia [zbitl] Fix and revamp CRC32 checking
The CRC-checking mode was completely broken.  It computed the
payload CRC regardless of whether the item header had the CRC flag
set.  It then never checked the CRC it computed against the header.
This went unnoticed due to the confluence of two issues.  First, the
test is permissive (by design), assuming any non-storage error is
the expected CRC mismatch error.  Second, the test data in the
corpus was bad, having a corrupted header instead of a corrupted
payload.  The test now checks for the exact error string, and the
test data is now correct.

The StorageTraits::Crc32 method is replaced with a generalized Read
method that takes a callback function.  This will be used to support
other library-based payload access like ZBI_FLAG_STORAGE_COMPRESSED
decompression in the future.

Original-Bug: 49438
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412774
Original-Revision: b6398c7ece475a3c8cff4bef4a1c55ffb5d9482a
GitOrigin-RevId: bbf94e73b37a377965551552fa9611d532340e86
Change-Id: I113248d78749c6569ced5a48ba169e9aeb7a75b3