[roll] Roll fuchsia [GN SDK] Using architecture specific host tools
This changes the references to host tools in the build rule templates
and scripts to use the architecture specific tools.

If the host_cpu variable in the build is not supported, the x64
version is used.

For the scripts in bin/ the architecture is determined by mapping
`uname -m` to x64 or arm64. If the mapping fails, the tool path used
is tools/$(uname -m).

Original-Bug: 56744

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412474
Original-Revision: 63f4404a64189dc40889f2ef33ded0a325dc3277
GitOrigin-RevId: 3d840744740933d8e7eab95a8fc0739c2794a1a2
Change-Id: Ib5acd84472bc83877e31ee4f7d412758fda518e0