[roll] Roll fuchsia [vfs] Change AlphabeticalTraversal -> TraversalPosition.
We shouldn't insist on servers having to return directory entries in
alphabetical order, so this change allows servers to choose how they
want to return entries.  At this time no implementation takes advantage
of this, but at least now it's possible in future.

Since I was touching it anyway, in this CL I have also changed the
read_dirents API in a way that I think is an improvement:

 1. Made it async.  AsyncReadDirents offers little benefit since it is
    trivial to write a synchronous version of an asynchronous function,
    and performance is unlikely to be an issue.

 2. There was no need for it to take Arc<Self> as a parameter since there
    should be no transfer in ownership required; &self should suffice.

 3. TraversalPosition is now returned explicitly as a return output
    which makes it symmetric with the arguments.  This simplifies the
    dirents_sink traits which don't need to know about the traversal
    position.

I made similar changes to the lazy implementation and changed it so that
instead of passing two callbacks, a trait is used instead which
simplifies the things somewhat.

These changes allowed then allowed me to fix the subtle bug in fatfs
where we would incorrectly handle listings of directories that have
entries that sort before ".".

Original-Bug: 53796
Original-Fixed: 57087
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412856
Original-Revision: 12775862bd1641f75e3e29dfca71b50963acaf01
GitOrigin-RevId: 7f938e86fbd710248b3cb85f02a839a6d06e9a10
Change-Id: Ie7c193612e280fe3876f2d1a6dff24995f40cd23