[roll] Roll fuchsia [usb-hub-rewrite] Initialize usb_ and bus_ prior to calling DdkAdd
In the normal DDK, DdkInit is invoked asynchronously. However,
in fake_ddk, DdkInit is invoked from within the call to DdkAdd.
This allows for the async loop to be initialized prior to Init completing,
resulting in a race when attempting to access bus_ or usb_.

Test: UBSAN in infra
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/413082
Original-Revision: 6ba6318ed1d725b06a8dc1baa42edc39ebcf329e
GitOrigin-RevId: 35bee165f9585a176c8f69568e9d7f95cd1da8a8
Change-Id: I0c75dda89815ec8434454ccba813e71691822e78