[roll] Roll fuchsia [wlan][sl4f] Fix locking around hanging get_update
Previously if a get_update hung from never getting an update from the
policy layer, the lock around the inner facade would still be held by
the hanging get_update even if the consumer timed out from the update.
After that, no other commands would work because they would be waiting
for the lock. Now the update listener is behind a separate lock and the
lock is behind a reference so that the lock and update stream can be
reset for each test.

Original-Bug: 56323
Test: act.py -c <config> -tc SavedNetworksTest
	tested that only the one test fails if it never gets an update
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407639
Original-Revision: 9a7a62e41984bf45016eab03994b5a314403d4af
GitOrigin-RevId: e1cffd02fe581712b534b7a2f59087fc375650a7
Change-Id: Ic9f216d5f91079ec0695390f9d0cd4d2dc9e9501