[roll] Roll fuchsia [acpi][x64] Remove AcpiWalker class.
Remove the AcpiWalker class and replace it with a call to a new
utility function which allows us to walk the ACPI namespace using a
lambda with capture storage for state instead of putting the state
into a class.

Also, get the PCI code out of the business of needing to care about
the internals of the AcpiWalker class.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/411940
Original-Revision: c1efb505744a9ed1a732ac35d4c9bb96c4031df5
GitOrigin-RevId: ec93df046731ad869fe3ce7ddddc4f8e18c69bad
Change-Id: I66fa1a7740897fd5056b3dbb661fcc8640534ad5