[roll] Roll fuchsia [brcmfmac] Use firmware provided reason code
In 57315, when firmware reported a DEAUTH_IND event, it provided reason
code 6 and 7. However, the reason code reported to wlanstack is 3. Turns
out it was created based on the event type.

The correct behavior seems to be forwarding the reason code directly
as-is.

Original-Bug: 57315
Test: Included

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412621
Original-Revision: 110f0cb27ab942142b8576b7ecfe624c24a45a3f
GitOrigin-RevId: 08fa60457eefa2d22fccb0c9247f4ced36852b25
Change-Id: Idd3a06e5f1ce3717845bba9800e0d74ea4ac75a3