[roll] Roll fuchsia [system-updater] Generate a stream of states
In preparation to yield status updates over FIDL when installing an
update, this change modifies the task that installs the update to yield
that data as it installs the update using a a State enum that closely
matches the fuchsia.update.installer.State union.

This change also modifies the state machine defined in the FIDL to
specify that the system updater always progresses through the
WaitToReboot state before entering the Reboot state.

As this change modifies the on-disk format of the update history, the
prior data will be lost when upgrading past this CL.  Otherwise, no
functional changes are expected.

Original-Bug: 55407
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410913
Original-Revision: 1e98c67f103a32d5e2f286a923efc6d3e1264a87
GitOrigin-RevId: 735024bb3209b5fa9856997f0072bbca54f6fb1c
Change-Id: Ice2810b48ce003250a28aaa2a9462b58c027693c