[roll] Roll fuchsia [acpi][x64] Break initial namespace walk into 2 passes.
Previously to this CL, the code which was walking the ACPI device tree
would look for an HDAS (Intel HDA audio controller device) device and
attempt to publish the NHLT as metadata bound to the controller's PCI
topological path for the device to pick up later on after the PCI bus
driver had published the controller's device.

The trouble here is that the PCI bus device was always going to be
encountered first, before the HDAS device (which is a child of the PCI
bus device).  This meant that the PCI bus device was being published,
allowing the driver to start initializing before the ACPI driver had a
chance to publish the HDAS NHLT metadata, and setting up a race to see
if the ACPI driver could publish the NHLT metadata before the HDA
driver loaded and went looking for it.

Now, the init phase has been broken into two passes.  During the first
pass, we handle any device quirks and publish any metadata.  During
the second pass, we actually start to publish devices

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/411941
Original-Revision: 901cae4e403e50e3d22cdf669571d223e8236c20
GitOrigin-RevId: 4ff0a7b3c674475c2c954d45fd9918d01e77b32d
Change-Id: Ia8818b184a4ce0658d879160bff169ad38adb4ed