[roll] Roll fuchsia [toolchain] Pass along library path for libLLVM
Without this change we are unable to build wasm32-unknown-unknown shared
libraries using our toolchain. For some reason that combination of
settings causes us to look for libLLVM-....so which is not in the path
for some reason.

This workaround unblocks us for now.

This CL also adds //src/diagnostics/lib/triage/wasm:wasm to
//src/diagnostics:tests to make sure wasm can build.

Original-Bug: 56076

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/406702
Original-Revision: 8bca749185315325d16e5ed2d3fc56868e8be88a
GitOrigin-RevId: eae72c7cfb414879c8c9b418c2abf971d5b8daa6
Change-Id: I2f497c36e058b3ed90c7fb7278b9eb3d52f611b7