[roll] Roll fuchsia [fidl] Change "resource" keyword to "resource_definition"
The "resource" keyword will be used for FTP-057 as a modifier on
structs, tables, and unions. But it is already used for defining new
kinds of resources, i.e. handle generalization. It is difficult to
overload the keyword for both usages because the parser is designed for
only one lookahead, so the easiest solution is to use different
keywords. Since the FTP-057 usage will be far more common than declaring
new kinds of resources, we prefer to reserve "resource" for the former,
and use the longer keyword "resource_definition" for the latter.

Test: fx test fidl-compiler --host
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412734
Original-Revision: fd67fa2373a6495967bc6b7aaed3b8970496e190
GitOrigin-RevId: 060b78de2ef46af0b9fa420262a12b20fea4f2de
Change-Id: I5cb3be9a9040b0b3f46441e2b338cb7a323bb8a0