[roll] Roll fuchsia [feedback] Add e2e test to verify inspect data in bugreports
This will prevent issues like the one seen this week. Through this
codepath the feedback component will be using the `core/archivist`
not an archivist injected to the enclosed environment as in the existing
integration tests.

The test is quite simple and can be extended to verify more data in
the future. At the moment it only checks for the health status of the
archivist in the inspect.json inside the bug report zip.

Tested: sl4f_client_tests, --e2e bugreport_test
Multiply: bugreport_test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/411845
Original-Revision: 1bd357e4b63695c684d7a856a5680da8574e7c53
GitOrigin-RevId: 3e1d2623fb753273a580a8aeb40c45d5aa287775
Change-Id: I92d8672519d0bf7fe35e5f00993c96854fc10061