[roll] Roll fuchsia [fidl][llcpp] refactor async requests
Use OwnedRequest to send the asynchronous request. That means that, now,
request <= 512 bytes are allocated on the stack.

Refactor dispatch. Now the handling of the contexts is generic. That
means that we no longer have a switch for the replies' ordinal (we only
have a switch for events).

A side effect is that fixes a vulnerability which could have been
exploted by a malicious server. With the previous implementation, a
reply was decoded using the ordinal within the message header then, the
result was used (with a cast) using the txid. An attacker could have
send a reply with an ordinal different from the request. This could have
crash the client or open a back door.

The asynchronous version is barely used. We only spare 1240 bytes.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/411924
Original-Revision: 00cd3e4248b502d6edfb99bc3fb7ae0a33ec38cb
GitOrigin-RevId: e86e253d18fca51dc84e77b1922390be538ef553
Change-Id: I45dcb7be8be0693da025140170af1d0571736656