[roll] Roll fuchsia [acpi][x86] Remove a potential read-past-end of buffer
Fix an issue where code which was extracting HID/CID information and
was attempting to pack it into a Fuchsia DDK devprops array. The
previous code would ignore a PNP Id string if it was more than 8 bytes
long, but it would _always_ read 8 bytes from the string array
regardless of how long the ACPICA library reported the string to be.
Now, we will read up to 8 bytes, but not past the end of the string
(packing 0s into the dev props if the string is too short.

Original-Fixed: 57531
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/411942
Original-Revision: 7c78b5e8c90f215c32049b858434f5089e25ed3e
GitOrigin-RevId: 4a7e35ea645d364bf5cd78991991ed2715eef1d0
Change-Id: I487f120a4f19daa5348da072ffdbf0dbd9b3d68a