[roll] Roll fuchsia [fdr] Show an overlay while counting down to reset
Show the usual overlay when counting down to a factory data reset. Use
F11 as volume down and F12 as volume up to test on a device without
consumer control keys.

Also added copy to the keyboard phase as there’s no reason for it not
to be copyable.

This is unlikely to be the last UI change to this program, but it is
worth committing as is.

Testing: manual
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412126
Original-Revision: 8b91b72f0e4d82f77f6fe26406845c02ad6a1a72
GitOrigin-RevId: a7009441ca298a021bf9a46bb93ada1bcc52fb0a
Change-Id: I0a54c34994cfdaf7d3e5d9015689a4b4767233f1