[roll] Roll fuchsia [dart] fix crash on malformed pubspec
The package importer is crashing because of a file with tabs
at the end of the line. This is not allowed in pubspec files
but we cannot skip the packages because they are required by
flutter. We trim the trailing tab and continue.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/413600
Original-Revision: 69b5173da82fb123a4136dfdf5e7b6de2b9d9eed
GitOrigin-RevId: 57e3755d381333a27e21bdda490338da073d3ca1
Change-Id: I497df344b7a7a8eb16fe132ae946085f0ea77494