[roll] Roll fuchsia [FDF] Exit gracefully when stop event completes
In component lifecycle stop event handler, driver manager suspends all drivers and
If the termination state is Reboot/shutdown, it also issues reboot/shutdown syscall.
At the end of the handler, closing the lifecycle handle causes the component manager
to kill driver manager. When driver manager which is marked critical is killed, the
root job watcher in kernel issues a reboot. If mexec is switched to use orderly shutdown,
this results in an undesired behavior where we want to mexec the system, not reboot.

This changeset exits the driver manager gracefully instead of closing
the lifecycle event handle. If we do so, kernel will not reboot the
system.

Test: fx reboot. fx netboot after moving mexec lib to orderly shutdown.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412222
Original-Revision: 558d78c1f730a0f81882dd6a90220824fa1c1a12
GitOrigin-RevId: ddb96f947810eb4892609f28e2f13bd3eff44e3f
Change-Id: I81dc7288d15f96e362ec11aeb2c0d551e5c6acf5