[roll] Roll fuchsia [time] Fix cobalt logging bug.
Although the first event was logging fine the CobaltMetrics object and
the Task it contained was being dropped as soon as time synchronization
succeeded, meaning the sucess event was never actually sent.

Longer term time sync will be a loop and so we won't have this problem,
but for now the easiest solution is just detach the task.

Original-Bug: 57677
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/413596
Original-Revision: b1a90f224c5f8e062b95b1598e61581bc5e3933a
GitOrigin-RevId: 7cf32435a036da732857473dbe0b64ad1cd64043
Change-Id: Icc306407eef60d05205404ad1bfc9699774c9589