[roll] Roll fuchsia [fuchsia-component] fail nested env creation with sub dirs
Nested environments are provided services from the root of a ServiceFs.
Developers are used to putting services inside of subdirectories. If a
developer creates a nested environment but puts the services for it in a
subdirectory, the nested environment will silently not receive these
services.

Update the create_nested_environment functions to check for sub
directories, and produce an error if they exist.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412655
Original-Revision: 5a38398466b19d82d1e547fb9e1f96727eea41e3
GitOrigin-RevId: 66eb46a6b89934c53270dc53f7a82fd1c2fc0ef6
Change-Id: I0ccf14715d6e625a02b0fedc2c85927b6c320779