[roll] Roll fuchsia [runtests] Remove --restrict-logs flag.
We are removing --restrict-logs flag from run-test-component, so need to
remove it from runtests. There is a alternative of --restrict-logs we
are developing but runtests will not support it as it is only used on
profile builders and there is no value added to fail the tests using
this/alternate flag on profile builders for now.

BUG=49735

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/411834
Original-Revision: 8aefbc0e41883e1c0ae241b57419435ae2a2c805
GitOrigin-RevId: 05ee46c192d46a8be30ce22926a8c2625872a966
Change-Id: Ifb4554827e96730b42a8929cb00e03fdb7db7e25