[roll] Roll fuchsia [SetUI] Store client/request with active requests
When sending a request through the registry, the original request
and the client should be stored so the events can be retried in case
of a failure (retries not in this CL).

This also fixes a bug in the regeneration test. A new handler is
created so that the event loop for the new request runs in the registry.

Add a note about how to prevent leaks when the retry behavior is
implemented.

Tests: fx test -o setui_service_tests
Original-Bug: 57168
Original-Fixed: 57166

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/413338
Original-Revision: 139e290844195ad7d882e1d16f0a53d12377bf37
GitOrigin-RevId: 99704dcdf4d558ecec867441fd20380da8bb176b
Change-Id: Ie5a7330cc0e99a18e7ffc1d7a7e6dbb4c62f4349