[roll] Roll fuchsia [aml-power][s905d2] Introduce power driver.
This patch introduces an power implementation driver for the amlogic
platform. Currently it is only implemented for the s905d2's CPU
power domain and currently only supports changing the voltage of the
power domain.

The functionality is duplicated from the aml-voltage component of the
aml-thermal driver which is ultimately planned to be removed from the
aml-thermal driver.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/401380
Original-Revision: 74800810780b7a691c1d504171b8c53d4e2f5bad
GitOrigin-RevId: 5277406a20a21fe94ad6c16f1862f4c94bc5d725
Change-Id: Ida5a60eb9dd7e66a57119ed2c9860852a3756743