[roll] Roll fuchsia [camera][bin] Connect FactoryServer to fuchsia.factory.camera.Isp
This CL changes the FactoryServer to pass incoming calls that target
the ISP or sensor to the corresponding calls on the ISP interface.

To necessitate this, the ISP's protodef was edited, and the device
was added to the component manifest.

A change was also made to the Controller protocol to simply pass a
channel to the ISP protocol instead of duplicating code.

This change also includes a fix for a linkage bug in the factory
server and an i2c write operation in the imx227 driver.

Test: visually confirmed that calls are relayed to the ISP by toggling
the sensor test patterns. This was done via an offline CL that attaches
a CLI to the camera-factory component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412745
Original-Revision: 632f4e32190eaff7c302ca92c7b0293b050cbac9
GitOrigin-RevId: 597905ab7bded630a33283efa4fdfaebfe4030a2
Change-Id: Icd6e8527ed1a4080c5d3e438cf93c347d4532e2e