[roll] Roll fuchsia [acpi][i2c] Publish I2C bus metadata from ACPI
Publish metadata describing the devices for I2C devices from ACPI.
This is a short term fix meant to remove the need for the PCI
protocol's GetAuxdata method.  Publishing the data from here will
allow the Intel I2C bus driver to fetch it's static metadata using the
DDK instead of a custom PCI protocol method.

In the longer term, publishing metadata from this level needs to stop
(both for I2C and NHLT).  It is simply not a scalable solution to need
to know a priori all of the meta data that all of the drivers are
going to need.  Instead we need to eventually move in a direction
where devices can interact with the ACPI driver directly at there
level of the ACPI device tree and below to fetch the metadata they are
interested in.

Original-Bug: 56832
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410446
Original-Revision: 303ad239981391fd5d0a1ca13e14a063c7688070
GitOrigin-RevId: 8fb279f198744fc45ce8c113b715621831d0745f
Change-Id: I39d52a565f86bce8b4d7bf76003615cf6f202da9