[roll] Roll fuchsia [validate_sysmgr] Print the correct conflict
This code previously referred to the `config_file` *loop* variable,
which always contained the path to the last file loaded, producing
errors like:

  Duplicate configuration for service fuchsia.net.http.Loader in files:
  ../../src/sys/timekeeper/service.config,
  ../../src/sys/timekeeper/service.config

which didn't make much sense. This change fixes that and uses
defaultdict to simplify the implementation.

It appears this bug was introduced in 94c35a1.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/413903
Original-Revision: 184db87182aee8578233623679095b06594b2f7b
GitOrigin-RevId: 3d2268feee8663a6935081120960af675a243cc6
Change-Id: I14ebce80281eba02239a3a1ee5696cd3dcafb4fb