[roll] Roll fuchsia [blackout] add ssh authentication configuration options
move off of fx ssh, and instead use regular ssh. in order for this to
work, we have to configure the ssh client ourselves in a sane way. this
adds a bunch of general configuration options to the ssh command, as
well as adding auth configuration for various situations.

Original-Fixed: 46959
Test: fx test //tools/blackout/blackout-host
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/367433
Original-Revision: 435b026a07ba2b5b36ab05ffef4e2f98b03d6506
GitOrigin-RevId: c0471ecd0f352eb16df75f559fd4abead7781c7d
Change-Id: I8bffb12a86195b0ad7c9f9905c839474176ed2a4