[roll] Roll fuchsia [Settings] Migrate back to watch
This a step in the process to rename Watch2 to Watch. A previous api
change required us to add a version of Watch that didn't return a
result/error, so it had to be renamed Watch2 so that both could
exist for a soft-transition. This change is to migrate back to the name
Watch.

Original-Bug: 55719, 55685
Test: fx test setui_service_tests cobalt_app_unittests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414001
Original-Revision: 379c2343579d55a07538ea39c567b132a2c8fec4
GitOrigin-RevId: 9b5238eefa9188c8c10557a591ddd70856ea11c7
Change-Id: I077c7506360c3066403bbe98a9daa41c43b4a496