[roll] Roll experiences [settings] Migrate back to watch
This a step in the process to rename Watch2 to Watch. A previous api
change required us to add a version of Watch that didn't return a
result/error, so it had to be renamed Watch2 so that both could
exist for a soft-transition. This change is to migrate back to the name
Watch.

Original-Bug: 55719
Test: fx run-host-tests ermine_settings_unittests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/414055
Original-Revision: ab8a8eeaa54b734036e95c0ab4507d72dd5d59ef
GitOrigin-RevId: a5d93044c819661b320162a8fe18da5e1531ad0a
Change-Id: Ie0f18134ef234e2d7edcb19a87b5c06ec6f300f6