[roll] Roll fuchsia [fidlc] Bound conditions on unknown enum member
When a flexible enum member is specified to be some value, allow the use
of the otherwise reserved max value.

Recording the unknown value of a flexible enum to later expose it in the
JSON IR. However, we first need to enable this feature in the compiler.

Test: added
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410875
Original-Revision: ee7b9022d0aebbfcc88ddfa89ccf09f2a08c154c
GitOrigin-RevId: 37e812d0f115ea02c3d31f524acce823a2815da8
Change-Id: I098cb5774d83e6f8c251776e253fe5c2fde8ecd7