[roll] Roll fuchsia [paver] add support for new partition scheme
Provides a standardized partition scheme that devices can use going
forward. Not all devices will be required to use this scheme, but for
cases that we have full control of the partition layout, we'll try to
keep as close to this as possible.

Additionally, this new scheme corrects two slight oddities with the old
strategy:
 * UUIDs were not generated randomly - several runs of bits were
   common between partitions
 * Partitions were uniquely identified by type GUID, whereas in the new
   scheme slotted partitions (e.g. zircon_{a,b,c}) will share the same
   type and be distinguished by name and unique GUID instead

Lastly, the debug names of partitions printed to the console have been
generalized instead of always using the legacy names. This fixes the
harmless but somewhat confusing cases like e.g. the Astro paver logging
that it's updating the "fuchsia-esp" partition which doesn't exist.

Original-Bug: 51200, 51442, 56775
Test: Sherlock pave
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409721
Original-Revision: 0010ded76edc81fcde8d4582461dc628447e3810
GitOrigin-RevId: caaf71ffe3cb513dc386d7b77dc0b9bf702dbd37
Change-Id: I84d1c39fced8d2d49b4243b27116c013f12e28ed