[roll] Roll fuchsia [crashsvc] Don't handle exceptions from exceptions.cmx
If exceptions.cmx crashes, crashsvc will attempt to send the exception
to the crashing process. Instead, if exceptions.cmx crashes terminate
the process by not handling its exception.

If sysmgr crashes early enough it could cause the same behavior, but
it is not being considered because there is no way to know if the
server for fuchsia.exception.Handler ever receives the other end of the
channel.

Original-Fixed: 56491

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/413891
Original-Revision: 4419fc0b1ff43885e6d513afa9d870720e0b200a
GitOrigin-RevId: 5ca26a8fc09487a3f433abf85605ca29b0bba225
Change-Id: I45d0045e8b50be4b64bc6ed6fe88ed4077fb95d9