[roll] Roll fuchsia [cml] Storage backing dir accepts name or path
Add a `backing_dir` field to the CML `storage` capability that accepts
the name of a directory capability. This will eventually replace `path`
which identifies the capability by `path`.

Putting this in a new field, instead of overloading `path`, simplifies
the implementation of cmc a bit. Also, it's closer to the final syntax
we'll end up with.

Note that the CM and CM-field still uses `storage_path` for both. In
this case, it turned out to be easier to reuse `storage_path` than to
add another field.

Original-Bug: 56604
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/413138
Original-Revision: 177b40de1e4004180eec4fcd0161ef73953b5c47
GitOrigin-RevId: eeb6e7dd2e0a9f5d666d3c6155e9ad8b09e02dce
Change-Id: I47690d92f639bf84487f101b07a1f5daf9a3a98c