[roll] Roll fuchsia [net-cli] Add dhcp start/stop
This change adds net-cli interfaces to start/stop a dhcp client on the
specified interface. This is part of an effort to bring interfaces that
currently only exist in ifconfig into net-cli, with the aim of
eventually deprecating and removing the former.

Original-Fixed: 56759

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414356
Original-Revision: 7afdcb35d57fec47c91a770ea30d5af9ea6d0784
GitOrigin-RevId: 88c883762b0e15fb68a929be140beba291d804b4
Change-Id: I2013c9dc546de513f6884dec6e228a559f1bface