[roll] Roll fuchsia [GN SDK] Adding absolute root path flag to cipd rule
This adds a flag to the cipd package rule template which causes the
root property of the cipd package definition to use an absolute path vs.
relative to root_build_dir.

This allows the cipd definition to be run outside the GN/ninja
environment, for example in a recipe.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414388
Original-Revision: 66eb623f489d3eb8733d85691c6409fdbbb38529
GitOrigin-RevId: a184af874785c3837af94a644d00ecc1c1c65937
Change-Id: I0c8fec95ea18f9545a6116225f33f9f793eba239