[roll] Roll fuchsia [fidlcat] fix summary
Displays all the sessions found for a given handle. This case is mostly
found within tests when a handle is sent to the same process.

Fixes SyscallEvent::GetValue to be able to also get outline fields.

Fixes display of message summary when the ordinal is not known.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412198
Original-Revision: e4eead127ac2cd853df273af476269fe986352eb
GitOrigin-RevId: ae37d23e446fb44857e7079ae2794cf9f09319ef
Change-Id: Id1280838896c96097a7a59ea42eb8901c2cddb92