[roll] Roll fuchsia [gn][rust] Massage source paths for verification
When a relative path to a generated source file is used in a Rust build
(include_str!), the depfile emitted by the compiler is relative to the
build directory but not necessarily canonical. For example,
'../../out/default/foo/bar'. This change makes sure that the paths
are massaged to something like 'foo/bar' (given that the build directory
is 'out/default').

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414457
Original-Revision: 72d9073f9fee2d178561d829b7493857ed55be58
GitOrigin-RevId: c3f2985410319f948e6bc9a8a2773d6bd131a353
Change-Id: Id18f725e2a21734cdc9e6fcca8a94f82cad178db