[roll] Roll fuchsia [time] Enable ZX_RIGHT_WAIT on test client.
Once this right is available on the userspace clock handle that is being
duplicated, this will let the test client announce when the clock
starts.

Note the CL also cleans up some of the terminology used for the various
clocks in the log comments, the original confused me a little during
debug.

Original-Bug: 56870
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414095
Original-Revision: c24416e74f2109778afbe6c1a3c072b089732a81
GitOrigin-RevId: 270c38ede03e66316a470ef09038df22314d7a31
Change-Id: Idc6ae8f1ac41beed5a849a4feac34fcef999c36e