[roll] Roll fuchsia [fidl] Combine visitor implementation for fidl_decode and fidl_validate
These are nearly identical, except for a few small details such as:
- Decoder mutates and therefore needs non-const pointers
- The validator only keeps track of handle count, not actual handles

As a result, the implementations can be combined.

The motivation for this is to reduce work when making changes to the
visitors - currently you need to do essentially identical changes
in both of these files.

Test: existing tests for these functions

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412456
Original-Revision: e18b1bb36dd5f17f5e0406f9112463d4faaa8c97
GitOrigin-RevId: dd9391c0df16fb4c04d352bc5c37217653ad34fd
Change-Id: Ib002af6e03f1fb55868296622f7d11ef43dedd6c