[roll] Roll fuchsia [llcpp] Host LLCPP conformance tests
Host tests are particularly helpful for LLCPP because it is used early
on in the system boot process. When working on benchmarks, I relied on
HLCPP host conformance tests because they used parts of LLCPP (decode),
but it would be useful to have the full thing.

This CL introduces a new optional flag to the fidl build rule
"host_llcpp", which enables building LLCPP on host. This exists because
only some LLCPP features are currently supported on host and building
all FIDL files is not possible.

Test: fx test --host fidl_llcpp_host_conformance_test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414097
Original-Revision: d65890865799421f03cc952664bb9fb2cdc7b6a8
GitOrigin-RevId: 90a917d24dfe28f9777af4e9f5a537d5f211bdaf
Change-Id: Iee9c93a596a095095ba2693e3c77fd1e2fe88d68