[roll] Roll fuchsia [testrunner] Record test outputs after each run.
Currently it runs all runs of a test and records all outputs of the
results at the end. However, that causes it to print the "ok"/"not ok"
status of each test run all at the end for a specific test so it's
difficult to see which test logs pertain to which test run.

Multiply: testrunner_tests

Original-Bug: 57814
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414011
Original-Revision: fa51944133ea7fae146793426c3d2f6a5c093e54
GitOrigin-RevId: b7f55289f0018bce5637b3bb0bb4b09ff2d0abd7
Change-Id: I086862c07f57adbd69203e004adcad4a952a0715