[roll] Roll fuchsia [zxdb] SCE #9: Introduce --forward option to continuing
Test: fuchsia-pkg://fuchsia.com/zxc-tests#meta/zxc_unittests.cmx
is a test with death assertions: running it under the debugger will
naturally yield to an exception. `continue` will try to resume the
thread to no effect, while `continue --forward` does indeed forward the
exception to the process-level handlers which resolve it; continuing in
this way, we can reach successfully finish the execution of the test.

Original-Bug: 48767
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412556
Original-Revision: fd3345338b5aba75d8cd06f49249910f1c28ac9c
GitOrigin-RevId: 821733705b360af24b8ad6f525d658e5317b6923
Change-Id: I6b19ba59b2380711bef90cbf509bd4f35d84da8a